@@ -1,4 +1,4 @@
-﻿@inject IJSRuntime JSRuntime
+@inject IJSRuntime JSRuntime
 @inject NavigationManager NavigationManager
 @implements IDisposable
 @code {
