@@ -1,4 +1,4 @@
-﻿@using AzureNamingTool.Models;
+@using AzureNamingTool.Models;
 <iframe src="@externalurl" style="height:600px;width:800px;"></iframe>
 
 @code {
