@@ -1,24 +1,25 @@
-﻿@using AzureNamingTool.Models;
+@using AzureNamingTool.Models;
 <p>
     The <span class="fw-bold">Admin Log</span> page allows the viewing of all the logged events within the Azure Naming Tool. This includes security events, configuration changes, and log activity. Use the <span class="fw-bold">Filter</span>  to filter the list of records.
 </p>
 <p>
     <img style="width:600px;" class="img-fluid" src="Screenshots/AdminLog1.png" alt="Admin Log" />
 </p>
-<div class="card">
-    <h5 class="card-header bg-default text-dark">
-        Page Sections/Actions
-    </h5>
-    <div class="card card-body @theme!.ThemeStyle">
-        <div class="card">
-            <div class="card-body @theme!.ThemeStyle">
+<div class="modern-card">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Page Sections/Actions</span>
+    </div>
+    <div class="modern-card-body">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <p>
                     <span class="fw-bold">Export Admin Log</span>
                 </p>
-                <p>This option allows you to export the <span class="fw-bold">Admin Log</span> to a CSV file.</p>        <div class="card">
-                </div>
+                <p>This option allows you to export the <span class="fw-bold">Admin Log</span> to a CSV file.</p>
+            </div>
         </div>
-            <div class="card-body @theme!.ThemeStyle">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <p>
                     <span class="fw-bold">Purge Admin Log</span>
                 </p>
