@@ -1,4 +1,4 @@
-﻿@using AzureNamingTool.Models;
+@using AzureNamingTool.Models;
 <p>
     When the site is first installed, the application will prompt for the <span class="fw-bold">Global Admin password</span> to be set. This password allows access to the configuration for the tool.
 </p>
