@@ -1,4 +1,4 @@
-﻿@using AzureNamingTool.Models;
+@using AzureNamingTool.Models;
 <p>
     The <span class="fw-bold">Configuration</span> page is used to view/set the configuration for the tool and the resource naming components. This page is separated into multiple sections. Each section displays the configuration for the individual components.
 </p>
@@ -16,16 +16,16 @@ else
 }
 @if (admin)
 {
-    <div class="card mb-3">
-        <h5 class="card-header bg-default text-dark">
-            Base Configuration
-        </h5>
-        <div class="card card-body @theme!.ThemeStyle">
+    <div class="modern-card mb-3">
+        <div class="modern-card-header">
+            <span class="modern-card-title">Base Configuration</span>
+        </div>
+        <div class="modern-card-body">
             <p>
                 This section provides the ability to configure the base components required for resource name generation.
             </p>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Components
                     </h5>
@@ -41,8 +41,8 @@ else
                     </p>
                 </div>
             </div>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Delimiters
                     </h5>
@@ -53,17 +53,17 @@ else
             </div>
         </div>
     </div>
-    <div class="card mb-3">
-        <h5 class="card-header bg-default text-dark">
-            Component Configuration
-        </h5>
-        <div class="card card-body @theme.ThemeStyle">
+    <div class="modern-card mb-3">
+        <div class="modern-card-header">
+            <span class="modern-card-title">Component Configuration</span>
+        </div>
+        <div class="modern-card-body">
             <p>
                 This section provides the ability to configure the individual components used in name generation. Each component provides the ability to add/edit/delete items, as well as set the desired sort order.
                 Note: The Resource Type and Location components allow limited configuration to ensure compatibility with Azure.
             </p>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Export
                     </h5>
@@ -72,8 +72,8 @@ else
                     </p>
                 </div>
             </div>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Refresh
                     </h5>
@@ -82,16 +82,16 @@ else
                     </p>
                 </div>
             </div>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Reset
                     </h5>
                     <p>When the tool is installed, a default configuration is used for each component. The <span class="fw-bold">Reset</span> feature allows the default configuration to be restored for the component.</p>
                 </div>
             </div>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Import
                     </h5>
@@ -100,24 +100,24 @@ else
             </div>
         </div>
     </div>
-    <div class="card mb-3">
-        <h5 class="card-header bg-default text-dark">
-            General Configuration
-        </h5>
-        <div class="card card-body @theme.ThemeStyle">
+    <div class="modern-card mb-3">
+        <div class="modern-card-header">
+            <span class="modern-card-title">General Configuration</span>
+        </div>
+        <div class="modern-card-body">
             <p>
                 This section provides functionality to modify the global configuration for the tool. Administrators can export the current configuration (combining all component configuration files into a JSON file), or import a new configuration. This feature is used for backup/restore of the installation configuration.
             </p>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Global Configuration
                     </h5>
                     <p>
                         This section allows the user to import/export a complete configuration for the Azure Naming Tool Application. All component configuration files are combined into a single JSON file. This file is used when backing up/restoring the tool configuration.
                     </p>
-                    @*<div class="card">
-                <div class="card-body">
+                    @*<div class="modern-card">
+                <div class="modern-card-body">
                 <h5>Policy</h5>
                 <p>
                 This section allows for the creation of an Azure Policy file. The configured components are converted to an Azure Policy Definition file to enforce the configuration with your subscription.
@@ -131,16 +131,16 @@ else
 }
 else
 {
-    <div class="card mb-3">
-        <h5 class="card-header bg-default text-dark">
-            Base Configuration
-        </h5>
-        <div class="card card-body @theme!.ThemeStyle">
+    <div class="modern-card mb-3">
+        <div class="modern-card-header">
+            <span class="modern-card-title">Base Configuration</span>
+        </div>
+        <div class="modern-card-body">
             <p>
                 This section displays the current required/optional base components for resource name generation.
             </p>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Components
                     </h5>
@@ -149,8 +149,8 @@ else
                     </p>
                 </div>
             </div>
-            <div class="card">
-                <div class="card-body @theme!.ThemeStyle">
+            <div class="modern-card mb-3">
+                <div class="modern-card-body">
                     <h5>
                         Delimiters
                     </h5>
@@ -161,11 +161,11 @@ else
             </div>
         </div>
     </div>
-    <div class="card mb-3">
-        <h5 class="card-header bg-default text-dark">
-            Component Configuration
-        </h5>
-        <div class="card card-body @theme.ThemeStyle">
+    <div class="modern-card mb-3">
+        <div class="modern-card-header">
+            <span class="modern-card-title">Component Configuration</span>
+        </div>
+        <div class="modern-card-body">
             <p>
                 This section displays the current required/optional components. Each component displays the current component options that can be selected.
             </p>
