@@ -1,17 +1,17 @@
-﻿@using AzureNamingTool.Models;
+@using AzureNamingTool.Models;
 <p>
     The <span class="fw-bold">Generate</span> page allows users to generate names with the selected options. Names can be generated for a single resource type or mutliple resource types. The <span class="fw-bold">Generate</span> page will display the required and optional components configured by the administrator.
 </p>
-<div class="card mb-3">
-    <h5 class="card-header bg-default text-dark">
-        Generate Mode
-    </h5>
-    <div class="card card-body @theme!.ThemeStyle">
+<div class="modern-card mb-3">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Generate Mode</span>
+    </div>
+    <div class="modern-card-body">
         <p>
             This section allows the user to select the desired resource type(s) name to generate. Use the <span class="fw-bold">Category</span> selector to filter the list of resource types.
         </p>
-        <div class="card">
-            <div class="card-body @theme!.ThemeStyle">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <h5>Single Resource Type</h5>
                 <p>This option will generate a single resource type name using the selected components. All naming guidelines will be applied.</p>
                 <p>
@@ -19,8 +19,8 @@
                 </p>
             </div>
         </div>
-        <div class="card">
-            <div class="card-body @theme!.ThemeStyle">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <h5>Multiple Resource Types</h5>
                 <p>This option will generate a resource type name for each selected resource type using the selected components. All naming guidelines will be applied. All components will be required.</p>
                 <p>
@@ -33,11 +33,11 @@
         </div>
     </div>
 </div>
-<div class="card mb-3">
-    <h5 class="card-header bg-default text-dark">
-        Generated Name(s)
-    </h5>
-    <div class="card card-body @theme.ThemeStyle">
+<div class="modern-card mb-3">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Generated Name(s)</span>
+    </div>
+    <div class="modern-card-body">
         <p>
             This section will display the generated name(s), along with any notes for the validation process. The list of names can be exported to a CSV file for use in other applications.
         </p>
@@ -46,31 +46,31 @@
         </p>
     </div>
 </div>
-<div class="card mb-3">
-    <h5 class="card-header bg-default text-dark">
-        Resource Type(s)
-    </h5>
-    <div class="card card-body @theme.ThemeStyle">
+<div class="modern-card mb-3">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Resource Type(s)</span>
+    </div>
+    <div class="modern-card-body">
         <p>
             This section allows the user to selected the desired resource type(s) name to generate. Use the <span class="fw-bold">Category</span> selector to filter the list of resource types.
         </p>
     </div>
 </div>
-<div class="card mb-3">
-    <h5 class="card-header bg-default text-dark">
-        Naming Guidelines (Single Resource Type)
-    </h5>
-    <div class="card card-body @theme.ThemeStyle">
+<div class="modern-card mb-3">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Naming Guidelines (Single Resource Type)</span>
+    </div>
+    <div class="modern-card-body">
         <p>
             This section displays the naming guidelines for the selected resource type. These guidelines wil be used to validate the generated name.
         </p>
     </div>
 </div>
-<div class="card mb-3">
-    <h5 class="card-header bg-default text-dark">
-        Components
-    </h5>
-    <div class="card card-body @theme.ThemeStyle">
+<div class="modern-card mb-3">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Components</span>
+    </div>
+    <div class="modern-card-body">
         <p>
             Select the desired values for the required components. Optional components may be omitted.
         </p>
