@@ -1,4 +1,4 @@
-﻿@using AzureNamingTool.Models;
+@using AzureNamingTool.Models;
 
 <p>
     The <span class="fw-bold">Reference</span> page displays all sample resource type names, using the configured components.
@@ -6,29 +6,29 @@
 <p>
     <img style="width:600px;" class="img-fluid" src="Screenshots/ReferencePage.png" alt="Reference Page" />
 </p>
-<div class="card">
-    <h5 class="card-header bg-default text-dark">
-        Page Sections/Actions
-    </h5>
-    <div class="card card-body @theme!.ThemeStyle">
-        <div class="card">
-            <div class="card-body @theme!.ThemeStyle">
+<div class="modern-card">
+    <div class="modern-card-header">
+        <span class="modern-card-title">Page Sections/Actions</span>
+    </div>
+    <div class="modern-card-body">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <h5>Example</h5>
                 <p>
                     This section displays a sample generated name for the resource type using the current components.
                 </p>
             </div>
         </div>
-        <div class="card">
-            <div class="card-body @theme!.ThemeStyle">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <h5>Components</h5>
                 <p>
                     This section displays the first option for each component. These components are used to generate the sample name.
                 </p>
             </div>
         </div>
-        <div class="card">
-            <div class="card-body @theme!.ThemeStyle">
+        <div class="modern-card mb-3">
+            <div class="modern-card-body">
                 <h5>Naming Guidelines</h5>
                 <p>
                     This section displays the naming guidelines for the resource type.
