@@ -1,6 +1,7 @@
-﻿@using AzureNamingTool.Helpers
+@using AzureNamingTool.Helpers
 @using AzureNamingTool.Models
 @using AzureNamingTool.Services
+@using AzureNamingTool.Services.Interfaces
 @using Microsoft.AspNetCore.Html
 @using System.Text.RegularExpressions
 @inject IJSRuntime JsRuntime
@@ -8,30 +9,34 @@
 @inject ILogger<EditModal> Logger
 @inject StateContainer state
 @inject ProtectedSessionStorage session
+@inject ServicesHelper ServicesHelper
 
-<div class="modal fade show d-block" tabindex="-1" role="dialog">
-    <div class="modal-backdrop fade show" @onclick="Cancel"></div>
-    <div class="modal-dialog modal-lg" style="z-index: 1050">
-        <!-- Pop it above the backdrop -->
-        <div class="modal-content @theme.ThemeStyle">
-            <div class="modal-header @headerstyle text-white">
-                <h3 class="modal-title">@title</h3>
+<div class="modern-modal-content">
+    <div class="modern-modal-body">
+        <div class="modern-card">
+            <div class="modern-card-header" style="@headerstyle">
+                <h3 class="modern-modal-title">@title</h3>
             </div>
-            <div class="modal-body">
-                <p>
-                    @((MarkupString)message)
-                </p>
-                @if (type == "ResourceType")
-                {
-                    if (Convert.ToBoolean(config.ResourceTypeEditingAllowed))
+            <div class="modern-card-body">
+                <div style="margin-bottom: 1rem;">
+                    <label for="itemName" class="modern-form-label">Name</label>
+                    <input title="Name" value="@itemName" type="text" class="modern-form-input" id="name" @onchange="@((ui) => { itemName = (string)ui.Value!;})" />
+                </div>
+                <div>
+                    <p>
+                        @((MarkupString)message)
+                    </p>
+                    @if (type == "ResourceType")
                     {
-                        <div class="alert alert-danger alert-dismissible fade show mb-3" role="alert">
+                        if (Convert.ToBoolean(config.ResourceTypeEditingAllowed))
+                    {
+                        <div class="alert alert-danger alert-dismissible fade show" style="margin-bottom: 1rem;" role="alert">
                             <h4>ATTENTION</h4>
-                            <div class="mb-3">
+                            <div style="margin-bottom: 1rem;">
                                 <p>
                                     Resource Type settings are configured by the Azure Naming Tool team and match the Azure portal validation requirements. Overriding these values may result in invalid names for a resource.
                                 </p>
-                                <p class="fw-bold">
+                                <p class="modern-font-semibold">
                                     Please see the <a href="https://github.com/mspnp/AzureNamingTool/wiki/Resource-Type-Editing" target="_blank">Resource Type Editing documentation</a> for guidance on metadata editing.
                                 </p>
                             </div>
@@ -44,9 +49,9 @@
                     }
                     else
                     {
-                        <div class="alert alert-warning alert-dismissible fade show mb-3" role="alert">
+                        <div class="alert alert-warning alert-dismissible fade show" style="margin-bottom: 1rem;" role="alert">
                             <h4>NOTE</h4>
-                            <div class="mb-3">
+                            <div style="margin-bottom: 1rem;">
                                 Resource Type settings are disabled by default. You can allow the editing of these values by enabling the "Resource Type Editing" setting in the Admin section.
                             </div>
                             <div>
@@ -57,263 +62,263 @@
                         </div>
                     }
                 }
-                <div class="card mb-3">
-                    <div class="card-header bg-default text-dark fw-bold">
+                <div class="modern-card" style="margin-bottom: 1rem;">
+                    <div class="modern-card-header">
                         Configuration
                     </div>
-                    <div class="card card-body @theme.ThemeStyle">
-                        <div class="mb-3">
+                    <div class="modern-card-body">
+                        <div style="margin-bottom: 1rem;">
                             This section allows you to modify the configuration.
                         </div>
-                        <div class="mb-3">
-                            <div class="fw-bold">
+                        <div style="margin-bottom: 1rem;">
+                            <label class="modern-form-label">
                                 @namelabel
-                            </div>
+                            </label>
                             <div>
                                 @if ((isProtectedName) || ((type == "ResourceType") && (!Convert.ToBoolean(config.ResourceTypeEditingAllowed))))
                                 {
-                                    <input title="Name" value="@itemName" disabled type="text" class="form-control @theme.ThemeStyle" id="name" />
+                                    <input title="Name" value="@itemName" disabled type="text" class="modern-form-input" id="name" />
                                 }
                                 else
                                 {
-                                    <input title="Name" value="@itemName" type="text" class="form-control @theme.ThemeStyle" id="name" @onchange="@((ui) => { itemName = (string)ui.Value!;})" />
+                                    <input title="Name" value="@itemName" type="text" class="modern-form-input" id="name" @onchange="@((ui) => { itemName = (string)ui.Value!;})" />
                                 }
                             </div>
                         </div>
                         @if (type != "ResourceComponent")
                         {
-                            <div class="mb-3">
-                                <div class="fw-bold">
+                            <div style="margin-bottom: 1rem;">
+                                <label class="modern-form-label">
                                     Short Name
-                                </div>
+                                </label>
                                 <div>
                                     @if (((type == "ResourceType") && (!Convert.ToBoolean(config.ResourceTypeEditingAllowed))))
                                     {
-                                        <input title="Short Name" disabled value="@itemShortName" type="text" class="form-control @theme.ThemeStyle" id="shortname" />
+                                        <input title="Short Name" disabled value="@itemShortName" type="text" class="modern-form-input" id="shortname" />
                                     }
                                     else
                                     {
-                                        <input title="Short Name" value="@itemShortName" type="text" class="form-control @theme.ThemeStyle" id="shortname" @onchange="@((ui) => { itemShortName = (string)ui.Value!;})" />
+                                        <input title="Short Name" value="@itemShortName" type="text" class="modern-form-input" id="shortname" @onchange="@((ui) => { itemShortName = (string)ui.Value!;})" />
                                     }
                                 </div>
                             </div>
                         }
                         @if ((type == "ResourceComponent") || ((type == "ResourceType") && (Convert.ToBoolean(config.ResourceTypeEditingAllowed))))
                         {
-                            <div class="mb-3">
-                                <div class="fw-bold">
+                            <div style="margin-bottom: 1rem;">
+                                <label class="modern-form-label">
                                     Description
-                                </div>
+                                </label>
                                 <div>
-                                    <input title="Description" value="@itemDescription" type="text" class="form-control @theme.ThemeStyle" id="maxlength" @onchange="@((ui) => { itemDescription = (string)ui.Value!;})" />
+                                    <input title="Description" value="@itemDescription" type="text" class="modern-form-input" id="maxlength" @onchange="@((ui) => { itemDescription = (string)ui.Value!;})" />
                                 </div>
                             </div>
                             @if ((type == "ResourceComponent") && (itemIsCustom))
                             {
-                                <div class="mb-3">
-                                    <div class="fw-bold">
+                                <div style="margin-bottom: 1rem;">
+                                    <label class="modern-form-label">
                                         Field Type
-                                    </div>
-                                    <select @bind="itemComponentType" class="form-control @theme.ThemeStyle">
+                                    </label>
+                                    <select @bind="itemComponentType" class="modern-form-select">
                                         <option value="standard">Standard - User selects an option from a pre-defined list</option>
                                         <option value="freetext">Free Form / Random</option>
                                     </select>
                                 </div>
                                 @if (itemComponentType == "freetext")
                                 {
-                                    <div class="card card-body @theme.ThemeStyle mb-3">
+                                    <div class="modern-card-body" style="margin-bottom: 1rem;">
                                         <div>
-                                            <div class="fw-bold">
+                                            <label class="modern-form-label">
                                                 Generate Random Value
-                                            </div>
-                                            <div class="mb-3">
+                                            </label>
+                                            <div style="margin-bottom: 1rem;">
                                                 By default, Free Form allows any value. If enabled, this setting to generate a random value for the component.
                                             </div>
                                             <div>
-                                                <label class="switch" title="Allow Random Value Generation">
+                                                <label class="modern-toggle-switch" title="Allow Random Value Generation">
                                                     <input type="checkbox" checked="@itemEnforceRandom" @oninput='args => ComponentSettingChanged(args, "enforcerandom")'>
-                                                    <span class="slider round"></span>
+                                                    <span class="modern-toggle-slider"></span>
                                                 </label>
-                                                <span class="align-text-top"> Enable</span>
+                                                <span class="align-text-top ms-2"> Enable</span>
                                             </div>
                                         </div>
                                         @if (itemEnforceRandom)
                                         {
-                                            <div class="mt-3">
-                                                <div class="fw-bold">
+                                            <div style="margin-top: 1rem;">
+                                                <label class="modern-form-label">
                                                     Generate Alphanumeric Values
-                                                </div>
-                                                <div class="mb-3">
+                                                </label>
+                                                <div style="margin-bottom: 1rem;">
                                                     If enabled, the tool will generate a random alphanumeric value for the component. If disabled, only letters will be generated.
                                                 </div>
                                                 <div>
-                                                    <label class="switch" title="Alphanumeric">
+                                                    <label class="modern-toggle-switch" title="Alphanumeric">
                                                         <input type="checkbox" checked="@itemAlphanumeric" @oninput='args => ComponentSettingChanged(args, "alphanumeric")'>
-                                                        <span class="slider round"></span>
+                                                        <span class="modern-toggle-slider"></span>
                                                     </label>
-                                                    <span class="align-text-top"> Enable</span>
+                                                    <span class="align-text-top ms-2"> Enable</span>
                                                 </div>
                                             </div>
                                         }
                                     </div>
                                 }
                             }
-                            <div class="mb-3">
-                                <div class="fw-bold">
+                            <div style="margin-bottom: 1rem;">
+                                <label class="modern-form-label">
                                     Minimum Length
-                                </div>
+                                </label>
                                 <div>
-                                    <input title="Minimum Length" value="@itemMinLength" type="text" class="form-control @theme.ThemeStyle" id="minlength" @onchange="@((ui) => { itemMinLength = (string)ui.Value!;})" onkeypress="return (event.charCode !=8 && event.charCode ==0 || (event.charCode >= 48 && event.charCode <= 57))" />
+                                    <input title="Minimum Length" value="@itemMinLength" type="text" class="modern-form-input" id="minlength" @onchange="@((ui) => { itemMinLength = (string)ui.Value!;})" onkeypress="return (event.charCode !=8 && event.charCode ==0 || (event.charCode >= 48 && event.charCode <= 57))" />
                                 </div>
                             </div>
-                            <div class="mb-3">
-                                <div class="fw-bold">
+                            <div style="margin-bottom: 1rem;">
+                                <label class="modern-form-label">
                                     Maximum Length
-                                </div>
+                                </label>
                                 <div>
-                                    <input title="Maximum Length" value="@itemMaxLength" type="text" class="form-control @theme.ThemeStyle" id="maxlength" @onchange="@((ui) => { itemMaxLength = (string)ui.Value!;})" onkeypress="return (event.charCode !=8 && event.charCode ==0 || (event.charCode >= 48 && event.charCode <= 57))" />
+                                    <input title="Maximum Length" value="@itemMaxLength" type="text" class="modern-form-input" id="maxlength" @onchange="@((ui) => { itemMaxLength = (string)ui.Value!;})" onkeypress="return (event.charCode !=8 && event.charCode ==0 || (event.charCode >= 48 && event.charCode <= 57))" />
                                 </div>
                             </div>
                             @if (type == "ResourceComponent")
                             {
-                                <div class="mb-3">
-                                    <div class="fw-bold">
+                                <div style="margin-bottom: 1rem;">
+                                    <label class="modern-form-label">
                                         Apply Delimiter Before
-                                    </div>
-                                    <div class="mb-3">
+                                    </label>
+                                    <div style="margin-bottom: 1rem;">
                                         If disabled, this setting will not add the current delimiter BEFORE the component value.
                                     </div>
                                     <div>
-                                        <label class="switch" title="Apply Delimiter">
+                                        <label class="modern-toggle-switch" title="Apply Delimiter">
                                             <input type="checkbox" checked="@itemApplyDelimiterBefore" @oninput='args => ComponentSettingChanged(args, "applydelimiterbefore")'>
-                                            <span class="slider round"></span>
+                                            <span class="modern-toggle-slider"></span>
                                         </label>
-                                        <span class="align-text-top"> Enable</span>
+                                        <span class="align-text-top ms-2"> Enable</span>
                                     </div>
                                 </div>
-                                <div class="mb-3">
-                                    <div class="fw-bold">
+                                <div style="margin-bottom: 1rem;">
+                                    <label class="modern-form-label">
                                         Apply Delimiter After
-                                    </div>
-                                    <div class="mb-3">
+                                    </label>
+                                    <div style="margin-bottom: 1rem;">
                                         If disabled, this setting will not add the current delimiter AFTER the component value.
                                     </div>
                                     <div>
-                                        <label class="switch" title="Apply Delimiter">
+                                        <label class="modern-toggle-switch" title="Apply Delimiter">
                                             <input type="checkbox" checked="@itemApplyDelimiterAfter" @oninput='args => ComponentSettingChanged(args, "applydelimiterafter")'>
-                                            <span class="slider round"></span>
+                                            <span class="modern-toggle-slider"></span>
                                         </label>
-                                        <span class="align-text-top"> Enable</span>
+                                        <span class="align-text-top ms-2"> Enable</span>
                                     </div>
                                 </div>
                             }
                             if (type == "ResourceType")
                             {
-                                <div class="mb-3">
-                                    <div class="fw-bold">
+                                <div style="margin-bottom: 1rem;">
+                                    <label class="modern-form-label">
                                         Apply Delimiter
-                                    </div>
-                                    <div class="mb-3">
+                                    </label>
+                                    <div style="margin-bottom: 1rem;">
                                         If disabled, the current delimiter will not be used for name generation.
                                     </div>
                                     <div>
-                                        <label class="switch" title="Apply Delimiter">
+                                        <label class="modern-toggle-switch" title="Apply Delimiter">
                                             <input type="checkbox" checked="@itemApplyDelimiter" @oninput='args => ComponentSettingChanged(args, "applydelimiter")'>
-                                            <span class="slider round"></span>
+                                            <span class="modern-toggle-slider"></span>
                                         </label>
-                                        <span class="align-text-top"> Enable</span>
+                                        <span class="align-text-top ms-2"> Enable</span>
                                     </div>
                                 </div>
-                                <div id="typemetdadatacontainer" class="card mb-3">
-                                    <div class="card-header bg-default text-dark fw-bold">
+                                <div id="typemetdadatacontainer" class="modern-card" style="margin-bottom: 1rem;">
+                                    <div class="modern-card-header">
                                         <a id="typemetdadatalink" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="#typemetdadata" role="button" aria-expanded="false" aria-controls="typemetdadata">
                                             <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Resoure Type Metadata
                                         </a>
                                     </div>
-                                    <div class="collapse card card-body @theme.ThemeStyle" id="typemetdadata">
-                                        <div class="mb-3">
+                                    <div class="collapse modern-card-body" id="typemetdadata">
+                                        <div style="margin-bottom: 1rem;">
                                             <p>
                                                 This section allows you to edit the Resouce Type metadata.
                                             </p>
-                                            <div class="alert alert-danger mb-3" role="alert">
+                                            <div class="alert alert-danger" style="margin-bottom: 1rem;" role="alert">
                                                 <h4>ATTENTION</h4>
-                                                <span class="fw-bold">Please see the <a href="https://github.com/mspnp/AzureNamingTool/wiki/Resource-Type-Editing" target="_blank">Resource Type Editing documentation</a> for guidance on metadata editing.</span>
+                                                <span class="modern-font-semibold">Please see the <a href="https://github.com/mspnp/AzureNamingTool/wiki/Resource-Type-Editing" target="_blank">Resource Type Editing documentation</a> for guidance on metadata editing.</span>
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="mb-3">
-                                                <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <div style="margin-bottom: 1rem;">
+                                                <label class="modern-form-label">
                                                     Regex
-                                                </div>
-                                                <div class="mb-3">
+                                                </label>
+                                                <div style="margin-bottom: 1rem;">
                                                     Regex to apply to generated names.
                                                 </div>
                                                 <div>
-                                                    <input title="RegEx" value="@itemRegEx" type="text" class="form-control @theme.ThemeStyle" id="regex" @onchange="@((ui) => { itemRegEx = (string)ui.Value!;})" />
+                                                    <input title="RegEx" value="@itemRegEx" type="text" class="modern-form-input" id="regex" @onchange="@((ui) => { itemRegEx = (string)ui.Value!;})" />
                                                 </div>
                                             </div>
 
-                                            <div class="fw-bold">
+                                            <label class="modern-form-label">
                                                 Scope
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Scope" value="@itemScope" type="text" class="form-control @theme.ThemeStyle" id="scope" @onchange="@((ui) => { itemScope = (string)ui.Value!;})" />
+                                                <input title="Scope" value="@itemScope" type="text" class="modern-form-input" id="scope" @onchange="@((ui) => { itemScope = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Valid Text
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Valid Text" value="@itemValidText" type="text" class="form-control @theme.ThemeStyle" id="validtext" @onchange="@((ui) => { itemValidText = (string)ui.Value!;})" />
+                                                <input title="Valid Text" value="@itemValidText" type="text" class="modern-form-input" id="validtext" @onchange="@((ui) => { itemValidText = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Text
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Text" value="@itemInvalidText" type="text" class="form-control @theme.ThemeStyle" id="invalidtext" @onchange="@((ui) => { itemInvalidText = (string)ui.Value!;})" />
+                                                <input title="Invalid Text" value="@itemInvalidText" type="text" class="modern-form-input" id="invalidtext" @onchange="@((ui) => { itemInvalidText = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters" value="@itemInvalidCharacters" type="text" class="form-control @theme.ThemeStyle" id="invalidcharacters" @onchange="@((ui) => { itemInvalidCharacters = (string)ui.Value!;})" />
+                                                <input title="Invalid Characters" value="@itemInvalidCharacters" type="text" class="modern-form-input" id="invalidcharacters" @onchange="@((ui) => { itemInvalidCharacters = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters Start
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters Start" value="@itemInvalidCharactersStart" type="text" class="form-control @theme.ThemeStyle" id="invalidcharactersstart" @onchange="@((ui) => { itemInvalidCharactersStart = (string)ui.Value!;})" />
+                                                <input title="Invalid Characters Start" value="@itemInvalidCharactersStart" type="text" class="modern-form-input" id="invalidcharactersstart" @onchange="@((ui) => { itemInvalidCharactersStart = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters End
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters End" value="@itemInvalidCharactersEnd" type="text" class="form-control @theme.ThemeStyle" id="invalidcharactersend" @onchange="@((ui) => { itemInvalidCharactersEnd = (string)ui.Value!;})" />
+                                                <input title="Invalid Characters End" value="@itemInvalidCharactersEnd" type="text" class="modern-form-input" id="invalidcharactersend" @onchange="@((ui) => { itemInvalidCharactersEnd = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters Consecutive
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters Consecutive" value="@itemInvalidCharactersConsecutive" type="text" class="form-control @theme.ThemeStyle" id="invalidcharactersconsecutive" @onchange="@((ui) => { itemInvalidCharactersConsecutive = (string)ui.Value!;})" />
+                                                <input title="Invalid Characters Consecutive" value="@itemInvalidCharactersConsecutive" type="text" class="modern-form-input" id="invalidcharactersconsecutive" @onchange="@((ui) => { itemInvalidCharactersConsecutive = (string)ui.Value!;})" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Static Values
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Static Values" value="@itemStaticValues" type="text" class="form-control @theme.ThemeStyle" id="staticvalues" @onchange="@((ui) => { itemStaticValues = (string)ui.Value!;})" />
+                                                <input title="Static Values" value="@itemStaticValues" type="text" class="modern-form-input" id="staticvalues" @onchange="@((ui) => { itemStaticValues = (string)ui.Value!;})" />
                                             </div>
                                         </div>
                                     </div>
@@ -324,11 +329,11 @@
                         {
                             @if ((type == "ResourceType") && (!Convert.ToBoolean(config.ResourceTypeEditingAllowed)))
                             {
-                                <div class="mb-3">
-                                    <div class="fw-bold">
+                                <div style="margin-bottom: 1rem;">
+                                    <label class="modern-form-label">
                                         Apply Delimiter
-                                    </div>
-                                    <div class="mb-3">
+                                    </label>
+                                    <div style="margin-bottom: 1rem;">
                                         If disabled, the current delimiter will not be used for name generation.
                                     </div>
                                     <div>
@@ -342,89 +347,89 @@
                                         }
                                     </div>
                                 </div>
-                                <div id="typemetdadatadisabledcontainer" class="card mb-3">
-                                    <div class="card-header bg-default text-dark fw-bold">
+                                <div id="typemetdadatadisabledcontainer" class="modern-card" style="margin-bottom: 1rem;">
+                                    <div class="modern-card-header">
                                         <a id="typemetdadatadisabledlink" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="#typemetdadatadisabled" role="button" aria-expanded="false" aria-controls="typemetdadatadisabled">
                                             <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Resoure Type Metadata
                                         </a>
                                     </div>
-                                    <div class="collapse card card-body @theme.ThemeStyle" id="typemetdadatadisabled">
-                                        <div class="mb-3">
+                                    <div class="collapse modern-card-body" id="typemetdadatadisabled">
+                                        <div style="margin-bottom: 1rem;">
                                             This section allows you to view the Resouce Type metadata.
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Regex
-                                            </div>
-                                            <div class="mb-3">
+                                            </label>
+                                            <div style="margin-bottom: 1rem;">
                                                 Regex to apply to generated names.
                                             </div>
                                             <div>
-                                                <input title="RegEx" value="@itemRegEx" type="text" disabled class="form-control @theme.ThemeStyle" id="regex" />
+                                                <input title="RegEx" value="@itemRegEx" type="text" disabled class="modern-form-input" id="regex" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Scope
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Scope" value="@itemScope" type="text" disabled class="form-control @theme.ThemeStyle" id="scope" />
+                                                <input title="Scope" value="@itemScope" type="text" disabled class="modern-form-input" id="scope" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Valid Text
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Valid Text" value="@itemValidText" type="text" disabled class="form-control @theme.ThemeStyle" id="validtext" />
+                                                <input title="Valid Text" value="@itemValidText" type="text" disabled class="modern-form-input" id="validtext" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Text
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Text" value="@itemInvalidText" type="text" disabled class="form-control @theme.ThemeStyle" id="invalidtext" />
+                                                <input title="Invalid Text" value="@itemInvalidText" type="text" disabled class="modern-form-input" id="invalidtext" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters" value="@itemInvalidCharacters" type="text" disabled class="form-control @theme.ThemeStyle" id="invalidcharacters" />
+                                                <input title="Invalid Characters" value="@itemInvalidCharacters" type="text" disabled class="modern-form-input" id="invalidcharacters" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters Start
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters Start" value="@itemInvalidCharactersStart" type="text" disabled class="form-control @theme.ThemeStyle" id="invalidcharactersstart" />
+                                                <input title="Invalid Characters Start" value="@itemInvalidCharactersStart" type="text" disabled class="modern-form-input" id="invalidcharactersstart" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters End
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters End" value="@itemInvalidCharactersEnd" type="text" disabled class="form-control @theme.ThemeStyle" id="invalidcharactersend" />
+                                                <input title="Invalid Characters End" value="@itemInvalidCharactersEnd" type="text" disabled class="modern-form-input" id="invalidcharactersend" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Invalid Characters Consecutive
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Invalid Characters Consecutive" value="@itemInvalidCharactersConsecutive" type="text" disabled class="form-control @theme.ThemeStyle" id="invalidcharactersconsecutive" />
+                                                <input title="Invalid Characters Consecutive" value="@itemInvalidCharactersConsecutive" type="text" disabled class="modern-form-input" id="invalidcharactersconsecutive" />
                                             </div>
                                         </div>
-                                        <div class="mb-3">
-                                            <div class="fw-bold">
+                                        <div style="margin-bottom: 1rem;">
+                                            <label class="modern-form-label">
                                                 Static Values
-                                            </div>
+                                            </label>
                                             <div>
-                                                <input title="Static Values" value="@itemStaticValues" type="text" disabled class="form-control @theme.ThemeStyle" id="staticvalues" />
+                                                <input title="Static Values" value="@itemStaticValues" type="text" disabled class="modern-form-input" id="staticvalues" />
                                             </div>
                                         </div>
                                     </div>
@@ -436,14 +441,14 @@
                 </div>
                 @if (type == "ResourceComponent")
                 {
-                    <div id="globallyoptionalcontainer" class="card mb-3">
-                        <div class="card-header bg-default text-dark fw-bold">
+                    <div id="globallyoptionalcontainer" class="modern-card" style="margin-bottom: 1rem;">
+                        <div class="modern-card-header">
                             <a id="OptionalComponents" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="#globallyoptional" role="button" aria-expanded="false" aria-controls="globallyoptional">
                                 <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Globally Optional Configuration
                             </a>
                         </div>
-                        <div class="collapse card card-body @theme.ThemeStyle" id="globallyoptional">
-                            <div class="mb-3">
+                        <div class="collapse modern-card-body" id="globallyoptional">
+                            <div style="margin-bottom: 1rem;">
                                 This section allows you to add/remove the component as OPTIONAL for all resource types.
                             </div>
                             <table class="table @theme.ThemeStyle">
@@ -453,7 +458,7 @@
                                             This option will ADD the component as OPTIONAL for all resource types.
                                         </td>
                                         <td>
-                                            <button type="button" class="btn btn-success w-100" @onclick="@(e => FormAction("ResourceComponent","optional-add"))" style="min-width:42px;" title="Add as OPTIONAL for all resoure types">
+                                            <button type="button" class="modern-btn-success w-100" @onclick="@(e => FormAction("ResourceComponent","optional-add"))" style="min-width:42px;" title="Add as OPTIONAL for all resoure types">
                                                 ADD
                                             </button>
                                         </td>
@@ -463,7 +468,7 @@
                                             This option will REMOVE the component from OPTIONAL for all resource types.
                                         </td>
                                         <td>
-                                            <button type="button" class="btn btn-danger w-100" @onclick="@(e => FormAction("ResourceComponent","optional-remove"))" style="min-width:42px;" title="Remove as OPTIONAL for all resoure types">
+                                            <button type="button" class="modern-btn-danger w-100" @onclick="@(e => FormAction("ResourceComponent","optional-remove"))" style="min-width:42px;" title="Remove as OPTIONAL for all resoure types">
                                                 REMOVE
                                             </button>
                                         </td>
@@ -472,14 +477,14 @@
                             </table>
                         </div>
                     </div>
-                    <div id="globallyexcludecontainer" class="card mb-3">
-                        <div class="card-header bg-default text-dark fw-bold">
+                    <div id="globallyexcludecontainer" class="modern-card" style="margin-bottom: 1rem;">
+                        <div class="modern-card-header">
                             <a id="OptionalComponents" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="#globallyexclude" role="button" aria-expanded="false" aria-controls="globallyexclude">
                                 <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Globally Exclude Configuration
                             </a>
                         </div>
-                        <div class="collapse card card-body @theme.ThemeStyle" id="globallyexclude">
-                            <div class="mb-3">
+                        <div class="collapse modern-card-body" id="globallyexclude">
+                            <div style="margin-bottom: 1rem;">
                                 This section allows you to add/remove the component as EXCLUDE for all resource types.
                             </div>
                             <table class="table">
@@ -489,7 +494,7 @@
                                             This option will ADD the component as EXCLUDE for all resource types.
                                         </td>
                                         <td>
-                                            <button type="button" class="btn btn-success w-100" @onclick="@(e => FormAction("ResourceComponent","exclude-add"))" style="min-width:42px;" title="Add as EXCLUDE for all resoure types">
+                                            <button type="button" class="modern-btn-success w-100" @onclick="@(e => FormAction("ResourceComponent","exclude-add"))" style="min-width:42px;" title="Add as EXCLUDE for all resoure types">
                                                 ADD
                                             </button>
                                         </td>
@@ -499,7 +504,7 @@
                                             This option will REMOVE the component from EXCLUDE for all resource types.
                                         </td>
                                         <td>
-                                            <button type="button" class="btn btn-danger w-100" @onclick="@(e => FormAction("ResourceComponent","exclude-remove"))" style="min-width:42px;" title="Remove as EXCLUDE for all resoure types">
+                                            <button type="button" class="modern-btn-danger w-100" @onclick="@(e => FormAction("ResourceComponent","exclude-remove"))" style="min-width:42px;" title="Remove as EXCLUDE for all resoure types">
                                                 REMOVE
                                             </button>
                                         </td>
@@ -511,13 +516,13 @@
                 }
                 @if ((type == "ResourceType") && (GeneralHelper.IsNotNull(resourceComponents)))
                 {
-                    <div id="optional" class="card mb-3">
-                        <div class="card-header bg-default text-dark fw-bold">
+                    <div id="optional" class="modern-card" style="margin-bottom: 1rem;">
+                        <div class="modern-card-header">
                             <a id="OptionalComponents" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="#optionalcomponents" role="button" aria-expanded="false" aria-controls="optionalcomponents">
                                 <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Optional Components
                             </a>
                         </div>
-                        <div class="collapse card card-body @theme.ThemeStyle" id="optionalcomponents">
+                        <div class="collapse modern-card-body" id="optionalcomponents">
                             @foreach (ResourceComponent resourceComponent in resourceComponents)
                             {
                                 if (resourceComponent.Name != "ResourceType")
@@ -543,13 +548,13 @@
                         </div>
                     </div>
 
-                    <div id="excluded" class="card mb-3">
-                        <div class="card-header bg-default text-dark fw-bold">
+                    <div id="excluded" class="modern-card" style="margin-bottom: 1rem;">
+                        <div class="modern-card-header">
                             <a id="ExcludedComponents" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="#excludedcomponents" role="button" aria-expanded="false" aria-controls="excludedcomponents">
                                 <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Excluded Components
                             </a>
                         </div>
-                        <div class="collapse card card-body @theme.ThemeStyle" id="excludedcomponents">
+                        <div class="collapse modern-card-body" id="excludedcomponents">
                             @foreach (ResourceComponent resourceComponent in resourceComponents)
                             {
                                 //@if (resourceComponent.Name != "ResourceType")
@@ -576,16 +581,15 @@
                     </div>
                 }
             </div>
-            <div class="modal-footer">
-
-                <button title="Save" @onclick="Save" class="btn btn-success">Update</button>
-
-                <button title="Cancel" @onclick="Cancel" class="btn btn-secondary">Close</button>
-            </div>
         </div>
     </div>
 </div>
+</div>
 
+<div class="modern-modal-footer">
+    <button title="Save" @onclick="Save" class="modern-btn-success">Update</button>
+    <button title="Cancel" @onclick="Cancel" class="modern-btn-secondary">Close</button>
+</div>
 
 
 @code {
@@ -603,6 +607,42 @@
     [Parameter] public string headerstyle { get; set; } = String.Empty;
     [Parameter] public bool admin { get; set; } = false;
     [Parameter] public ServicesData servicesData { get; set; } = new();
+    
+    [Inject]
+    public IResourceComponentService? ResourceComponentService { get; set; }
+    
+    [Inject]
+    public IAdminUserService? AdminUserService { get; set; }
+    
+    [Inject]
+    public IResourceEnvironmentService? ResourceEnvironmentService { get; set; }
+    
+    [Inject]
+    public IResourceLocationService? ResourceLocationService { get; set; }
+    
+    [Inject]
+    public IResourceOrgService? ResourceOrgService { get; set; }
+    
+    [Inject]
+    public IResourceProjAppSvcService? ResourceProjAppSvcService { get; set; }
+    
+    [Inject]
+    public IResourceTypeService? ResourceTypeService { get; set; }
+    
+    [Inject]
+    public IResourceUnitDeptService? ResourceUnitDeptService { get; set; }
+    
+    [Inject]
+    public IResourceFunctionService? ResourceFunctionService { get; set; }
+    
+    [Inject]
+    public IResourceDelimiterService? ResourceDelimiterService { get; set; }
+    
+    [Inject]
+    public ICustomComponentService? CustomComponentService { get; set; }
+    
+    [Inject]
+    public IAdminLogService? AdminLogService { get; set; }
 
     private ServiceResponse serviceResponse = new();
     private string itemName = String.Empty;
@@ -690,11 +730,12 @@
                 itemShortName = resourceProjAppSvc.ShortName;
                 break;
             case "ResourceType":
-                serviceResponse = await ResourceTypeService.GetItem(id);
+                serviceResponse = await ResourceTypeService.GetItemAsync(id);
                 ResourceType resourceType = servicesData.ResourceTypes!.Find(x => x.Id == id)!;
                 itemName = resourceType.Resource;
                 namelabel = "Resource";
                 itemShortName = resourceType.ShortName;
+                itemDescription = resourceType.Description ?? String.Empty;
                 optional = resourceType.Optional;
                 exclude = resourceType.Exclude;
                 itemRegEx = resourceType.Regx;
@@ -733,7 +774,7 @@
         StateHasChanged();
     }
 
-    async void Save()
+    async Task Save()
     {
         try
         {
@@ -760,7 +801,7 @@
                         foreach (var customcomponent in customcomponents)
                         {
                             customcomponent.ParentComponent = GeneralHelper.NormalizeName(itemName, true);
-                            serviceResponse = await CustomComponentService.PostItem(customcomponent);
+                            serviceResponse = await CustomComponentService.PostItemAsync(customcomponent);
                         }
                         resourceComponent.Name = itemName;
                         if (resourceComponent.IsCustom)
@@ -791,11 +832,11 @@
                         if ((itemOriginalStandard) && (resourceComponent.IsFreeText))
                         {
                             bool confirm = false;
-                            confirm = await ModalHelper.ShowConfirmationModal(Modal!, "ATTENTION", "<div class=\"my-4\">This will update the custom component to free text and remove the current component options! This cannot be un-done!</div><div class=\"my-4\">Are you sure?</div>", "bg-danger", theme);
+                            confirm = await ModalHelper.ShowConfirmationModal(Modal!, "ATTENTION", "<div style='margin: 1.5rem 0;'>This will update the custom component to free text and remove the current component options! This cannot be un-done!</div><div style='margin: 1.5rem 0;'>Are you sure?</div>", "bg-danger", theme);
                             if (confirm)
                             {
-                                serviceResponse = await CustomComponentService.DeleteByParentComponentId((int)resourceComponent.Id);
-                                serviceResponse = await ResourceComponentService.PostItem(resourceComponent);
+                                serviceResponse = await CustomComponentService.DeleteByParentComponentIdAsync((int)resourceComponent.Id);
+                                serviceResponse = await ResourceComponentService.PostItemAsync(resourceComponent);
                             }
                             else
                             {
@@ -804,7 +845,7 @@
                         }
                         else
                         {
-                            serviceResponse = await ResourceComponentService.PostItem(resourceComponent);
+                            serviceResponse = await ResourceComponentService.PostItemAsync(resourceComponent);
                         }
                     }
                     else
@@ -821,7 +862,7 @@
             {
                 if ((!String.IsNullOrEmpty(itemName)) && (!String.IsNullOrEmpty(itemShortName)))
                 {
-                    if (await ValidationHelper.ValidateShortName(type, itemShortName, this.parentcomponent))
+                    if (await ValidationHelper.ValidateShortName(ResourceComponentService!, type, itemShortName, this.parentcomponent))
                     {
                         try
                         {
@@ -840,28 +881,29 @@
                                 ResourceEnvironment resourceEnvironment = servicesData.ResourceEnvironments!.Find(x => x.Id == id)!;
                                 resourceEnvironment.Name = itemName;
                                 resourceEnvironment.ShortName = itemShortName;
-                                serviceResponse = await ResourceEnvironmentService.PostItem(resourceEnvironment);
+                                serviceResponse = await ResourceEnvironmentService.PostItemAsync(resourceEnvironment);
                                 break;
                             case "ResourceLocation":
                                 ResourceLocation resourceLocation = servicesData.ResourceLocations!.Find(x => x.Id == id)!;
                                 resourceLocation.ShortName = itemShortName;
-                                serviceResponse = await ResourceLocationService.PostItem(resourceLocation);
+                                serviceResponse = await ResourceLocationService.PostItemAsync(resourceLocation);
                                 break;
                             case "ResourceOrg":
                                 ResourceOrg resourceOrg = servicesData.ResourceOrgs!.Find(x => x.Id == id)!;
                                 resourceOrg.Name = itemName;
                                 resourceOrg.ShortName = itemShortName;
-                                serviceResponse = await ResourceOrgService.PostItem(resourceOrg);
+                                serviceResponse = await ResourceOrgService.PostItemAsync(resourceOrg);
                                 break;
                             case "ResourceProjAppSvc":
                                 ResourceProjAppSvc resourceProjAppSvc = servicesData.ResourceProjAppSvcs!.Find(x => x.Id == id)!;
                                 resourceProjAppSvc.Name = itemName;
                                 resourceProjAppSvc.ShortName = itemShortName;
-                                serviceResponse = await ResourceProjAppSvcService.PostItem(resourceProjAppSvc);
+                                serviceResponse = await ResourceProjAppSvcService.PostItemAsync(resourceProjAppSvc);
                                 break;
                             case "ResourceType":
                                 ResourceType resourceType = servicesData.ResourceTypes!.Find(x => x.Id == id)!;
                                 resourceType.ShortName = itemShortName;
+                                resourceType.Description = itemDescription;
                                 resourceType.Optional = "";
                                 resourceType.Exclude = "";
                                 resourceType.ApplyDelimiter = itemApplyDelimiter;
@@ -921,25 +963,25 @@
                                         }
                                     }
                                 }
-                                serviceResponse = await ResourceTypeService.PostItem(resourceType);
+                                serviceResponse = await ResourceTypeService.PostItemAsync(resourceType);
                                 break;
                             case "ResourceUnitDept":
                                 ResourceUnitDept resourceUnitDept = servicesData.ResourceUnitDepts!.Find(x => x.Id == id)!;
                                 resourceUnitDept.Name = itemName;
                                 resourceUnitDept.ShortName = itemShortName;
-                                serviceResponse = await ResourceUnitDeptService.PostItem(resourceUnitDept);
+                                serviceResponse = await ResourceUnitDeptService.PostItemAsync(resourceUnitDept);
                                 break;
                             case "ResourceFunction":
                                 ResourceFunction resourceFunction = servicesData.ResourceFunctions!.Find(x => x.Id == id)!;
                                 resourceFunction.Name = itemName;
                                 resourceFunction.ShortName = itemShortName;
-                                serviceResponse = await ResourceFunctionService.PostItem(resourceFunction);
+                                serviceResponse = await ResourceFunctionService.PostItemAsync(resourceFunction);
                                 break;
                             case "CustomComponent":
                                 CustomComponent customComponent = servicesData.CustomComponents!.Find(x => x.Id == id)!;
                                 customComponent.Name = itemName;
                                 customComponent.ShortName = itemShortName;
-                                serviceResponse = await CustomComponentService.PostItem(customComponent);
+                                serviceResponse = await CustomComponentService.PostItemAsync(customComponent);
                                 break;
                         }
                     }
@@ -960,7 +1002,7 @@
                 {
                     await ModalInstance.CloseAsync();
                     toastService.ShowSuccess(GeneralHelper.NormalizeName(type, false) + " updated!");
-                    AdminLogService.PostItem(new AdminLogMessage() { Title = "SUCCESS", Message = "(" + GeneralHelper.NormalizeName(type, false) + ") " + itemName + " updated!", Source = currentuser });
+                    await AdminLogService?.PostItemAsync(new AdminLogMessage() { Title = "SUCCESS", Message = "(" + GeneralHelper.NormalizeName(type, false) + ") " + itemName + " updated!", Source = currentuser });
                 }
                 else
                 {
@@ -970,7 +1012,7 @@
         }
         catch (Exception ex)
         {
-            AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+            await AdminLogService?.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
             toastService.ShowError("There was an error updating the " + GeneralHelper.NormalizeName(type, false) + "! " + ex.Message);
         }
     }
@@ -1014,28 +1056,28 @@
                                     switch (action)
                                     {
                                         case "optional-add":
-                                            serviceResponse = await ResourceTypeService.UpdateTypeComponents("optional-add", componentid);
+                                            serviceResponse = await ResourceTypeService.UpdateTypeComponentsAsync("optional-add", componentid);
                                             if (serviceResponse.Success)
                                             {
                                                 responseMessage.Message += " ADDED as OPTIONAL!";
                                             }
                                             break;
                                         case "optional-remove":
-                                            serviceResponse = await ResourceTypeService.UpdateTypeComponents("optional-remove", componentid);
+                                            serviceResponse = await ResourceTypeService.UpdateTypeComponentsAsync("optional-remove", componentid);
                                             if (serviceResponse.Success)
                                             {
                                                 responseMessage.Message += " REMOVED from OPTIONAL!";
                                             }
                                             break;
                                         case "exclude-add":
-                                            serviceResponse = await ResourceTypeService.UpdateTypeComponents("exclude-add", componentid);
+                                            serviceResponse = await ResourceTypeService.UpdateTypeComponentsAsync("exclude-add", componentid);
                                             if (serviceResponse.Success)
                                             {
                                                 responseMessage.Message += " ADDED as EXCLUDE!";
                                             }
                                             break;
                                         case "exclude-remove":
-                                            serviceResponse = await ResourceTypeService.UpdateTypeComponents("exclude-remove", componentid);
+                                            serviceResponse = await ResourceTypeService.UpdateTypeComponentsAsync("exclude-remove", componentid);
                                             if (serviceResponse.Success)
                                             {
                                                 responseMessage.Message += " REMOVED from EXCLUDE!";
@@ -1049,7 +1091,7 @@
                     }
                     catch (Exception ex)
                     {
-                        AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                        await AdminLogService?.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                         responseMessage.Type = MessageTypesEnum.ERROR;
                         responseMessage.Message = "There was a problem with the request.";
                     }
@@ -1074,13 +1116,13 @@
                             toastService.ShowError(responseMessage.Message);
                             break;
                     }
-                    AdminLogService.PostItem(new AdminLogMessage() { Title = responseMessage.Type.ToString(), Message = responseMessage.Message, Source = currentuser });
+                    await AdminLogService?.PostItemAsync(new AdminLogMessage() { Title = responseMessage.Type.ToString(), Message = responseMessage.Message, Source = currentuser });
                 }
             }
         }
         catch (Exception ex)
         {
-            AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+            await AdminLogService?.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
             toastService.ShowError("There was an error with the action! " + ex.Message);
         }
     }
@@ -1102,7 +1144,7 @@
         }
     }
 
-    private void ComponentSettingChanged(ChangeEventArgs e, string setting)
+    private async Task ComponentSettingChanged(ChangeEventArgs e, string setting)
     {
         try
         {
@@ -1134,7 +1176,7 @@
         }
         catch (Exception ex)
         {
-            AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+            await AdminLogService?.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
             toastService.ShowError("There was an error updating the " + GeneralHelper.NormalizeName(type, false) + "! " + ex.Message);
         }
     }
@@ -1142,3 +1184,6 @@
 
     async Task Cancel() => await ModalInstance.CancelAsync();
 }
+
+
+
