@@ -1,95 +1,81 @@
-﻿@using AzureNamingTool.Helpers
+@using AzureNamingTool.Helpers
 @using AzureNamingTool.Models
 @using AzureNamingTool.Services
+@using AzureNamingTool.Services.Interfaces
 @using Microsoft.AspNetCore.Html
 @inject IToastService toastService
 @inject StateContainer state
 @inject IJSRuntime JsRuntime
+@inject ServicesHelper ServicesHelper
 
-<div class="modal fade show d-block" tabindex="-1" role="dialog">
-    <div class="modal-backdrop fade show" @onclick="Save"></div>
-    <div class="modal-dialog" style="z-index: 1050; max-width: 700px;">
-        <!-- Pop it above the backdrop -->
-        <div class="modal-content @theme.ThemeStyle">
-            <div class="modal-header bg-navcolor text-white">
-                <h3 class="modal-title">Select Resource Types</h3>
-            </div>
-            <div class="modal-body">
-                <p>
-                    Select the resource types.
-                </p>
+<div class="blazored-modal-container">
+    <div class="blazored-modal blazored-modal-large">
+        <div class="modern-modal-content">
+            <div class="modern-modal-body">
+                <div class="modern-card">
+                    <div class="modern-card-header">
+                        <h3 class="modern-card-title mb-0">Select Resource Types</h3>
+                    </div>
+                    <div class="modern-card-body">
+                        <p style="margin-bottom: 1rem;">
+                            Select the resource types.
+                        </p>
 
-                <div class="form-group row mt-3">
-                    <div class="input-group flex-nowrap">
-                        <div class="input-group flex-nowrap mb-3">
-                            <span class="input-group-text" id="addon-wrapping">
-                                <span class="oi oi-magnifying-glass" id="addon-wrapping"></span>
+                        <div class="modern-input-group mb-3">
+                            <span class="modern-input-group-icon">
+                                <span class="oi oi-magnifying-glass"></span>
                             </span>
-                            <input class="form-control" type="search" placeholder="Filter by Name, Short Name"
+                            <input class="modern-form-input" type="search" placeholder="Filter by Name, Short Name"
                                    @bind="filterData"
                                    @bind:event="oninput">
                         </div>
-                    </div>
-                </div>
 
-                <div class=" card @theme.ThemeStyle mb-3">
-                    <h5 class="card-header text-white bg-secondary text-white">
-                        <a id="ResourceDelimiter" class="link-light text-decoration-none" data-bs-toggle="collapse" style="display:block;" href="#resourcetypes" role="button" aria-expanded="false" aria-controls="resourcetypes">
-                            <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Resource Types
-                        </a>
-                    </h5>
-                    <div class="collapse show card card-body @theme.ThemeStyle" id="resourcetypes">
                         @foreach (var category in categories)
                         {
                             if (!IsCategoryVisible(category))
                                 continue;
-                            <div id="@("multitypecontainer_"+@category)" class="card mb-3">
-                                <div class="card-header bg-default text-dark">
-                                    <a id="NamingGuidlines" class="link-dark text-decoration-none" data-bs-toggle="collapse" href="@("#multitypesection_"+@category)" role="button" aria-expanded="false" aria-controls="@("multitypesection_"+@category)">
-                                        <span class="oi oi-chevron-bottom" aria-hidden="true"></span> @category
-                                    </a>
+                            <div class="modern-card mb-3">
+                                <div class="modern-card-header">
+                                    <h5 class="modern-card-title mb-0">@category</h5>
                                 </div>
-                                <div class="collapse show card card-body @theme.ThemeStyle" id="@("multitypesection_"+@category)">
-                                    <ul style="list-style-type: none;word-break: break-word;">
+                                <div class="modern-card-body">
+                                    <div class="modern-checkbox-group">
                                         @foreach (var resourceType in ResourceTypeService.GetFilteredResourceTypes(servicesData.ResourceTypes!, category))
                                         {
                                             if (!IsTypeVisible(resourceType))
                                                 continue;
-                                            <li>
-                                                <div class="form-check">
-                                                    @if (multiselectedResourceTypes.Contains(resourceType.Id.ToString()))
-                                                    {
-                                                        <input class="form-check-input" type="checkbox" checked value="@resourceType.Id" id="@("multitype_"+@resourceType.Id)" @onchange="eventArgs => { OnMultiTypesChange(resourceType.Id.ToString(), eventArgs.Value!); }">
-                                                    }
-                                                    else
-                                                    {
-                                                        <input class="form-check-input" type="checkbox" value="@resourceType.Id" id="@("multitype_"+@resourceType.Id)" @onchange="eventArgs => { OnMultiTypesChange(resourceType.Id.ToString(), eventArgs.Value!); }">
-                                                    }
+                                            <label class="modern-checkbox-label">
+                                                @if (multiselectedResourceTypes.Contains(resourceType.Id.ToString()))
+                                                {
+                                                    <input class="modern-checkbox" type="checkbox" checked value="@resourceType.Id" id="@("multitype_"+@resourceType.Id)" @onchange="eventArgs => { OnMultiTypesChange(resourceType.Id.ToString(), eventArgs.Value!); }">
+                                                }
+                                                else
+                                                {
+                                                    <input class="modern-checkbox" type="checkbox" value="@resourceType.Id" id="@("multitype_"+@resourceType.Id)" @onchange="eventArgs => { OnMultiTypesChange(resourceType.Id.ToString(), eventArgs.Value!); }">
+                                                }
+                                                <span class="modern-checkbox-text">
                                                     @if (!String.IsNullOrEmpty(resourceType.Property))
                                                     {
-                                                        <label class="form-check-label" for="flexCheckDefault">
-                                                            @resourceType.Resource - @resourceType.Property (@resourceType.ShortName)
-                                                        </label>
+                                                        <span>@resourceType.Resource - @resourceType.Property (@resourceType.ShortName)</span>
                                                     }
                                                     else
                                                     {
-                                                        <label class="form-check-label" for="flexCheckDefault">
-                                                            @resourceType.Resource (@resourceType.ShortName)
-                                                        </label>
+                                                        <span>@resourceType.Resource (@resourceType.ShortName)</span>
                                                     }
-                                                </div>
-                                            </li>
+                                                </span>
+                                            </label>
                                         }
-                                    </ul>
+                                    </div>
                                 </div>
                             </div>
                         }
                     </div>
                 </div>
             </div>
-            <div class="modal-footer">
-                <button title="Close" @onclick="Save" class="btn btn-success">Close</button>
-            </div>
+        </div>
+
+        <div class="modern-modal-footer">
+            <button title="Close" @onclick="Save" class="modern-btn-success">Close</button>
         </div>
     </div>
 </div>
@@ -100,6 +86,12 @@
     [Parameter] public string title { get; set; } = String.Empty;
     [Parameter] public string message { get; set; } = String.Empty;
     [Parameter] public List<string> multiselectedResourceTypes { get; set; } = new();
+    
+    [Inject]
+    public IResourceTypeService? ResourceTypeService { get; set; }
+    
+    [Inject]
+    public IAdminLogService? AdminLogService { get; set; }
 
     private ServicesData servicesData = new();
     private List<string> categories = new();
@@ -143,7 +135,7 @@
         }
         catch (Exception ex)
         {
-            AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+            await AdminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
             toastService.ShowError("There was a problem loading the page. Please try again");
         }
     }
@@ -167,7 +159,7 @@
         StateHasChanged();
     }
 
-    async void Save()
+    async Task Save()
     {
         await ModalInstance.CloseAsync(ModalResult.Ok(multiselectedResourceTypes));
     }
