@@ -1,40 +1,35 @@
-﻿@using AzureNamingTool.Helpers
+@using AzureNamingTool.Helpers
 @using AzureNamingTool.Models;
 @inject StateContainer state
 @inject ProtectedSessionStorage session
 @inject IToastService toastService
 @inject ILogger<PasswordModal> Logger
 
-<div class="modal fade show" id="myModal" style="display:block; background-color: rgba(10,10,10,.8);"
-     aria-modal="true" role="dialog">
-    <div class="modal-dialog modal-danger">
-        <div class="modal-content @state.AppTheme">
-            <div class="modal-body">
-                <div class="card">
-                    <div class="card-header bg-danger text-white">
-                        <h3>Global Admin Password Not Set!</h3>
-                    </div>
-                    <div class="card-body">
-                        <p>
-                            You must set the Global Admin Password to continue.
-                        </p>
-                        <p><span class="fw-bold">Requirements</span></p>
-                        <ul>
-                            <li>Contain a number</li>
-                            <li>Contain one upper case letter</li>
-                            <li>Be at least 8 characters</li>
-                        </ul>
-                        <div>
-                            <input id="newpassword" class="form-control mb-3" type="password" @onchange="@((ui) => { SetPassword((string?)ui.Value!);})" />
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="modal-footer">
-                <button type="button" disabled="@disabled" class="btn btn-primary" @onclick=ModalSave>Save</button>
-            </div>
+<div class="modern-modal-content">
+    <h3 style="margin-bottom: 1rem;"><i class="bi bi-shield-lock"></i> Set Global Admin Password</h3>
+    
+    <div class="modern-notification-warning mb-3">
+        <i class="bi bi-exclamation-triangle"></i>
+        <div>
+            <p>
+                You must set the Global Admin Password to continue.
+            </p>
+            <p class="modern-font-semibold">Requirements</p>
+            <ul>
+                <li>Contain a number</li>
+                <li>Contain one upper case letter</li>
+                <li>Be at least 8 characters</li>
+            </ul>
         </div>
     </div>
+    <div>
+        <label class="modern-form-label">New Password</label>
+        <input id="newpassword" class="modern-form-input mb-3" type="password" @onchange="@((ui) => { SetPassword((string?)ui.Value!);})" placeholder="Enter password" />
+    </div>
+</div>
+
+<div class="modern-modal-footer">
+    <button type="button" disabled="@disabled" class="modern-btn-primary" @onclick=ModalSave>Save</button>
 </div>
 
 @code {
@@ -44,12 +39,10 @@
     [Parameter]
     public EventCallback<bool> OnClose { get; set; }
 
-    private string password = String.Empty;
+    [CascadingParameter]
+    BlazoredModalInstance ModalInstance { get; set; } = new();
 
-    private Task ModalCancel()
-    {
-        return OnClose.InvokeAsync(false);
-    }
+    private string password = String.Empty;
 
     private async Task ModalSave()
     {
@@ -82,6 +75,7 @@
                     base.StateHasChanged();
                 });
 
+                await ModalInstance.CloseAsync(ModalResult.Ok(true));
                 await OnClose.InvokeAsync(true);
                 toastService.ShowSuccess("The Global Admin Password has been set.");
             }
