@@ -1,46 +1,34 @@
-﻿@using AzureNamingTool.Helpers
+@using AzureNamingTool.Helpers
 @using AzureNamingTool.Models;
 @inject StateContainer state
 @inject ProtectedSessionStorage session
 @inject IToastService toastService
 @inject ILogger<PasswordModal> Logger
 
-<div class="modal fade show" id="myModal" style="display:block; background-color: rgba(10,10,10,.8);"
-     aria-modal="true" role="dialog">
-    <div class="modal-backdrop fade show" @onclick="Cancel"></div>
-    <div class="modal-dialog" style="z-index: 1050">
-        <div class="modal-content @state.AppTheme">
-            <div class="modal-body mx-auto">
-                <div class="p-5 text-center">
-                    @if (!String.IsNullOrEmpty(message))
-                    {
-                        <p class="p-3" style="font-style:italic;">
-                            @message
-                        </p>
-                    }
-                    else
-                    {
-                        <p class="p-3" style="font-style:italic;">
-                            Working....
-                        </p>
-                    }
-                    <img class="p-3" src="/images/loading-buffering.gif" />
-                </div>
-            </div>
-        </div>
-    </div>
+<div class="modern-modal-content text-center" style="align-items: center; justify-content: center;">
+    @if (!String.IsNullOrEmpty(message))
+    {
+        <p style="margin-bottom: 1rem; font-style:italic; font-size: 1.1rem;">
+            @message
+        </p>
+    }
+    else
+    {
+        <p style="margin-bottom: 1rem; font-style:italic; font-size: 1.1rem;">
+            Working....
+        </p>
+    }
+    <img src="/images/loading-buffering.gif" alt="Loading..." style="max-width: 100px; display: block; margin: 0 auto;" />
 </div>
 
 @code {
     [Parameter] public EventCallback<bool> OnClose { get; set; }
+    [CascadingParameter] BlazoredModalInstance ModalInstance { get; set; } = new();
     [Parameter] public string message { get; set; } = String.Empty;
 
-    private Task Cancel()
+    private async Task Cancel()
     {
-        return OnClose.InvokeAsync(false);
-    }
-
-    private void ModalSave()
-    {
+        await ModalInstance.CancelAsync();
+        await OnClose.InvokeAsync(false);
     }
 }
