@@ -1,4 +1,4 @@
-﻿@page "/404"
+@page "/404"
 <img src="/images/runnoft.png" />
 <p role="alert">Whoa, it looks like that page doesn't exist! Please check the URL and try again!</p>
 @code {
