@@ -1,7 +1,8 @@
-﻿@page "/generatednameslog"
+@page "/generatednameslog"
 @using AzureNamingTool.Helpers
 @using AzureNamingTool.Models
 @using AzureNamingTool.Services
+@using AzureNamingTool.Services.Interfaces
 @using System.IO
 @using BlazorDownloadFile
 @using Microsoft.AspNetCore.Components.Authorization
@@ -18,25 +19,29 @@
 @inject NavigationManager NavigationManager
 
 <PageTitle>Azure Naming Tool - Generated Names Log</PageTitle>
-<div class="card  @theme.ThemeStyle">
-    <div class="card-body">
-        <div style="float:left;">
-            <h1>
-                Generated Names Log
-            </h1>
-        </div>
-        <div style="float:right;">
-            <button class="btn btn-secondary" title="Instructions" @onclick="@(e => ModalHelper.ShowInformationModal(Modal!, theme, "bg-navcolor", "Generated Names Log", String.Empty, "generatednamesloginstructions", false))">
-                <span class="oi oi-document" aria-hidden="true"></span> Documentation
-            </button>
-        </div>
-        <div style="clear:both;"></div>
-        <div class="mb-3">This page displays a log of generated resouce type names.</div>
 
+<div class="modern-page-container">
+    <div class="modern-page-title">
+        <h1>Generated Names Log</h1>
+    </div>
+    
+    <div class="modern-page-description">
+        <p>View a log of generated resource type names</p>
+    </div>
+    
+    <div class="modern-page-actions">
+        <button class="modern-btn-secondary" title="Instructions" @onclick="@(e => ModalHelper.ShowInformationModal(Modal!, theme, "bg-navcolor", "Generated Names Log", String.Empty, "generatednamesloginstructions", false))">
+            <span class="oi oi-document" aria-hidden="true"></span> Documentation
+        </button>
+    </div>
+</div>
+
+<div class="modern-card">
+    <div class="modern-card-body">
         @if (!dataLoaded)
         {
-            <div class="spinner-border blue" role="status">
-                <span class="sr-only"></span>
+            <div class="modern-spinner-container">
+                <div class="modern-spinner"></div>
             </div>
         }
         else
@@ -44,73 +49,78 @@
             <div class="mb-3">
                 @if (lstNameRequests.Count > 0)
                 {
-                    <div class="mb-3">
-                        <button class="btn btn-primary" title="Export Generated Names Log" @onclick="@(e => OnExportLogClick())">
+                    <div class="modern-button-group mb-3">
+                        <button class="modern-btn-primary" title="Export Generated Names Log" @onclick="@(e => OnExportLogClick())">
+                            <span class="oi oi-data-transfer-download" aria-hidden="true"></span>
                             Export Generated Names Log
                         </button>
                         @if (admin)
                         {
                             if (duplicatenamesallowed)
                             {
-                                <button class="btn btn-danger" style="margin-left:5px;" title="Purge Generated Names Log" @onclick="@(e => OnPurgeLogClick())">
+                                <button class="modern-btn-danger" title="Purge Generated Names Log" @onclick="@(e => OnPurgeLogClick())">
+                                    <span class="oi oi-trash" aria-hidden="true"></span>
                                     Purge Generated Names Log
                                 </button>
                             }
                             else
                             {
-                                <button class="btn btn-danger" style="margin-left:5px;" title="Purge Generated Names Log" disabled>
+                                <button class="modern-btn-danger" title="Purge Generated Names Log" disabled>
+                                    <span class="oi oi-trash" aria-hidden="true"></span>
                                     Purge Generated Names Log
                                 </button>
-                                <div class="mt-3">
-                                    <div class="alert alert-warning">
-                                        <span class="fw-bold">NOTE</span>:
-                                        Duplicate names are not currently allowed in the site settings. Purging of the log is restricted.
-                                    </div>
-                                </div>
                             }
                         }
                     </div>
-                    <div class="card mb-3">
-                        <div class="card-header bg-default text-dark">
-                            <a class="text-decoration-none text-dark" data-bs-toggle="collapse" style="display:block;font-weight:bold;" href="#filter" role="button" aria-expanded="false" aria-controls="filter">
-                                <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Filter Data
-                            </a>
+                    @if (admin && !duplicatenamesallowed)
+                    {
+                        <div class="modern-alert modern-alert-warning mb-3">
+                            <div>
+                                <strong>NOTE:</strong> Duplicate names are not currently allowed in the site settings. Purging of the log is restricted.
+                            </div>
                         </div>
-                        <div class="collapse show card card-body @theme.ThemeStyle" id="filter">
+                    }
+                    <div class="modern-card mb-3">
+                        <div class="modern-card-header collapsible collapsed" data-bs-toggle="collapse" data-bs-target="#filter" aria-expanded="false">
+                            <h3>
+                                <span class="oi oi-chevron-bottom" aria-hidden="true"></span> Filter Data
+                            </h3>
+                        </div>
+                        <div class="collapse modern-card-body" id="filter">
                             <div class="form-group row mb-3">
                                 <div style="width:auto;">
-                                    <input class="form-control" type="date" style="width: 150px;" @bind="startdate" @bind:event="oninput" />
+                                    <label class="modern-form-label">Start Date</label>
+                                    <input class="modern-form-input" type="date" style="width: 150px;" @bind="startdate" @bind:event="oninput" />
                                 </div>
-                                <div style="width:auto;padding-top:7px;">
+                                <div style="width:auto;padding-top:35px;padding-left:10px;padding-right:10px;">
                                     to
                                 </div>
                                 <div style="width:auto;">
-                                    <input class="form-control" type="date" style="width: 150px;" @bind="enddate" @bind:event="oninput" />
+                                    <label class="modern-form-label">End Date</label>
+                                    <input class="modern-form-input" type="date" style="width: 150px;" @bind="enddate" @bind:event="oninput" />
                                 </div>
                             </div>
-                            <div class="form-group row mb-3">
-                                <div class="input-group flex-nowrap">
-                                    <div class="input-group flex-nowrap">
-                                        <span class="input-group-text" id="addon-wrapping">
-                                            <span class="oi oi-magnifying-glass" id="addon-wrapping"></span>
-                                        </span>
-                                        <input class="form-control" type="search" placeholder="Filter by Created By, Generated Name, Resource Type, Components"
-                                        @bind="filterData"
-                                        @bind:event="oninput">
-                                    </div>
+                            <div class="form-group mb-3">
+                                <label class="modern-form-label">Filter by Created By, Generated Name, Resource Type, Components</label>
+                                <div class="modern-input-group">
+                                    <span class="modern-input-group-icon">
+                                        <span class="oi oi-magnifying-glass" aria-hidden="true" title="Search"></span>
+                                    </span>
+                                    <input class="modern-form-input modern-input-with-icon" type="search" placeholder="Filter data..."
+                                    @bind="filterData"
+                                    @bind:event="oninput">
                                 </div>
                             </div>
-                            <div class="form-group row">
-                                <div>
-                                    <button class="btn btn-success" title="Reset" @onclick="@(e => OnResetClick())">
-                                        Reset
-                                    </button>
-                                </div>
+                            <div class="form-group">
+                                <button class="modern-btn-success" title="Reset" @onclick="@(e => OnResetClick())">
+                                    <span class="oi oi-reload" aria-hidden="true"></span>
+                                    Reset
+                                </button>
                             </div>
                         </div>
                     </div>
-                    <div class="table-responsive">
-                        <table class="table @theme.ThemeStyle">
+                    <div class="modern-table-container">
+                        <table class="modern-table">
                             <thead>
                                 <tr>
                                     @if ((admin) && (duplicatenamesallowed))
@@ -134,7 +144,7 @@
                                         @if ((admin) && (duplicatenamesallowed))
                                         {
                                             <td class="align-top">
-                                                <button type="button" class="btn btn-danger" @onclick="@(e => DeleteGeneratedName(name.Id))" title="Delete">
+                                                <button type="button" class="modern-btn-danger modern-btn-sm" @onclick="@(e => DeleteGeneratedName(name.Id))" title="Delete">
                                                     <span class="oi oi-x" aria-hidden="true"></span>
                                                 </button>
                                             </td>
@@ -177,9 +187,10 @@
                 }
                 else
                 {
-                    <p style="font-style:italic">
-                        There are no log items to display.
-                    </p>
+                    <div class="modern-empty-state">
+                        <span class="oi oi-info" aria-hidden="true"></span>
+                        <p>There are no log items to display.</p>
+                    </div>
                 }
             </div>
         }
@@ -191,6 +202,13 @@
     protected ThemeInfo theme { get; set; } = new();
     [CascadingParameter]
     public IModalService? Modal { get; set; }
+    
+    [Inject]
+    public IGeneratedNamesService? GeneratedNamesService { get; set; }
+    
+    [Inject]
+    public IAdminLogService? AdminLogService { get; set; }
+    
     private bool admin = false;
     private bool dataLoaded = false;
     private bool duplicatenamesallowed = false;
@@ -222,7 +240,7 @@
     private async Task LoadData()
     {
         ServiceResponse serviceResponse = new();
-        serviceResponse = await GeneratedNamesService.GetItems();
+        serviceResponse = await GeneratedNamesService.GetItemsAsync();
         if (serviceResponse.Success)
         {
             if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
@@ -251,10 +269,10 @@
         ServiceResponse serviceResponse = new();
         ResponseMessage message = new();
         bool confirm = false;
-        confirm = await ModalHelper.ShowConfirmationModal(Modal!, "ATTENTION", "<div class=\"my-4\">This will delete the Generated Name and cannot be undone!</div><div class=\"my-4\">Are you sure?</div>", "bg-danger", theme);
+        confirm = await ModalHelper.ShowConfirmationModal(Modal!, "ATTENTION", "<div style='margin: 1.5rem 0;'>This will delete the Generated Name and cannot be undone!</div><div style='margin: 1.5rem 0;'>Are you sure?</div>", "bg-danger", theme);
         if (confirm)
         {
-            serviceResponse = await GeneratedNamesService.GetItem(Convert.ToInt32(id));
+            serviceResponse = await GeneratedNamesService.GetItemAsync(Convert.ToInt32(id));
             if (serviceResponse.Success)
             {
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
@@ -262,13 +280,13 @@
                     GeneratedName generatedname = (GeneratedName)serviceResponse.ResponseObject!;
                     if (GeneralHelper.IsNotNull(generatedname))
                     {
-                        serviceResponse = await GeneratedNamesService.DeleteItem(Convert.ToInt32(generatedname.Id));
+                        serviceResponse = await GeneratedNamesService.DeleteItemAsync(Convert.ToInt32(generatedname.Id));
                         if (serviceResponse.Success)
                         {
                             message.Type = MessageTypesEnum.SUCCESS;
                             message.Message = "Generated name deleted!";
                             toastService.ShowSuccess(message.Message);
-                            AdminLogService.PostItem(new AdminLogMessage() { Source = currentuser, Title = "INFORMATION", Message = "Generated Name (" + generatedname.ResourceName + ") deleted." });
+                            await AdminLogService.PostItemAsync(new AdminLogMessage() { Source = currentuser, Title = "INFORMATION", Message = "Generated Name (" + generatedname.ResourceName + ") deleted." });
                             CacheHelper.InvalidateCacheObject("GeneratedName");
                             await LoadData();
                             state.SetNavReload(true);
@@ -305,13 +323,22 @@
         ServiceResponse serviceResponse = new();
 
         bool confirm = false;
-        confirm = await ModalHelper.ShowConfirmationModal(Modal!, "ATTENTION", "<div class=\"my-4\">This will purge the Generated Names Log and cannot be undone!</div><div class=\"my-4\">Are you sure?</div>", "bg-danger", theme);
+        confirm = await ModalHelper.ShowConfirmationModal(Modal!, "ATTENTION", "<div style='margin: 1.5rem 0;'>This will purge the Generated Names Log and cannot be undone!</div><div style='margin: 1.5rem 0;'>Are you sure?</div>", "bg-danger", theme);
         if (confirm)
         {
-            await LogHelper.PurgeGeneratedNames();
-            toastService.ShowSuccess("Generated Names Log purged!");
-            AdminLogService.PostItem(new AdminLogMessage() { Source = currentuser, Title = "INFORMATION", Message = "Generated Names log purged." });
-            await LoadData();
+            // Use GeneratedNamesService.DeleteAllItemsAsync() which works with both SQLite and FileSystem
+            serviceResponse = await GeneratedNamesService!.DeleteAllItemsAsync();
+            
+            if (serviceResponse.Success)
+            {
+                toastService.ShowSuccess("Generated Names Log purged!");
+                await LoadData();
+            }
+            else
+            {
+                toastService.ShowError("Failed to purge Generated Names Log. Please check the Admin Log.");
+            }
+            
             StateHasChanged();
         }
     }
