@@ -1,4 +1,4 @@
-﻿@using AzureNamingTool.Models
+@using AzureNamingTool.Models
 
 <CascadingValue Value="identityProviderDetails">
     <CascadingBlazoredModal>
