@@ -1,12 +1,15 @@
-﻿using AzureNamingTool.Models;
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+using AzureNamingTool.Models;
 using AzureNamingTool.Helpers;
 using Microsoft.AspNetCore.Mvc;
+using Asp.Versioning;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
 using AzureNamingTool.Services;
+using AzureNamingTool.Services.Interfaces;
 using AzureNamingTool.Attributes;
 using Microsoft.AspNetCore.DataProtection.KeyManagement;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
@@ -17,11 +20,27 @@ namespace AzureNamingTool.Controllers
     /// Controller for managing Admin settings.
     /// </summary>
     [Route("api/[controller]")]
+    [ApiVersion("1.0")]
     [ApiController]
     [ApiKey]
+    [Produces("application/json")]
     public class AdminController : ControllerBase
     {
+        private readonly IAdminService _adminService;
+        private readonly IAdminLogService _adminLogService;
+        private readonly IGeneratedNamesService _generatedNamesService;
         private readonly SiteConfiguration config = ConfigurationHelper.GetConfigurationData();
+
+        public AdminController(
+            IAdminService adminService,
+            IAdminLogService adminLogService,
+            IGeneratedNamesService generatedNamesService)
+        {
+            _adminService = adminService;
+            _adminLogService = adminLogService;
+            _generatedNamesService = generatedNamesService;
+        }
+
         private ServiceResponse serviceResponse = new();
 
         // POST api/<AdminController>
@@ -33,6 +52,7 @@ namespace AzureNamingTool.Controllers
         /// <returns>string - Successful update</returns>
         [HttpPost]
         [Route("[action]")]
+        [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
         public async Task<IActionResult> UpdatePassword([BindRequired][FromHeader(Name = "AdminPassword")] string adminpassword, [FromBody] string password)
         {
             try
@@ -41,7 +61,7 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.UpdatePassword(password);
+                        serviceResponse = await _adminService.UpdatePasswordAsync(password);
                         return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem updating the password."));
                     }
                     else
@@ -57,7 +77,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -79,8 +99,8 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.UpdateAPIKey(apikey, "fullaccess");
-                        return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem updating the Full Access API Key."));
+                        serviceResponse = await _adminService.UpdateAPIKeyAsync(apikey, "fullaccess");
+                        return (serviceResponse.Success ? Ok("SUCCESS") : BadRequest(" - There was a problem updating the Full Access API Key."));
                     }
                     else
                     {
@@ -95,7 +115,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -117,8 +137,8 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.GenerateAPIKey("fullaccess");
-                        return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem generating the Full Access API Key."));
+                        serviceResponse = await _adminService.GenerateAPIKeyAsync("fullaccess");
+                        return (serviceResponse.Success ? Ok("SUCCESS") : BadRequest(" - There was a problem generating the Full Access API Key."));
                     }
                     else
                     {
@@ -133,7 +153,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -158,8 +178,8 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.UpdateAPIKey(apikey, "namegeneration");
-                        return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem updating the Name Generation API Key."));
+                        serviceResponse = await _adminService.UpdateAPIKeyAsync(apikey, "namegeneration");
+                        return (serviceResponse.Success ? Ok("SUCCESS") : BadRequest(" - There was a problem updating the Name Generation API Key."));
                     }
                     else
                     {
@@ -174,7 +194,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -195,8 +215,8 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.GenerateAPIKey("namegeneration");
-                        return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem generating the Name Generation API Key."));
+                        serviceResponse = await _adminService.GenerateAPIKeyAsync("namegeneration");
+                        return (serviceResponse.Success ? Ok("SUCCESS") : BadRequest(" - There was a problem generating the Name Generation API Key."));
                     }
                     else
                     {
@@ -211,7 +231,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -233,8 +253,8 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.UpdateAPIKey(apikey, "readonly");
-                        return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem updating the Read-Only API Key."));
+                        serviceResponse = await _adminService.UpdateAPIKeyAsync(apikey, "readonly");
+                        return (serviceResponse.Success ? Ok("SUCCESS") : BadRequest(" - There was a problem updating the Read-Only API Key."));
                     }
                     else
                     {
@@ -249,7 +269,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -270,8 +290,8 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminService.GenerateAPIKey("readonly");
-                        return (serviceResponse.Success ? Ok("SUCCESS") : Ok("FAILURE - There was a problem generating the Read-Only API Key."));
+                        serviceResponse = await _adminService.GenerateAPIKeyAsync("readonly");
+                        return (serviceResponse.Success ? Ok("SUCCESS") : BadRequest(" - There was a problem generating the Read-Only API Key."));
                     }
                     else
                     {
@@ -286,7 +306,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -305,7 +325,7 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminLogService.GetItems();
+                        serviceResponse = await _adminLogService.GetItemsAsync();
                         if (serviceResponse.Success)
                         {
                             return Ok(serviceResponse.ResponseObject);
@@ -328,7 +348,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -347,7 +367,7 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await AdminLogService.DeleteAllItems();
+                        serviceResponse = await _adminLogService.DeleteAllItemsAsync();
                         if (serviceResponse.Success)
                         {
                             return Ok(serviceResponse.ResponseObject);
@@ -370,7 +390,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -385,7 +405,7 @@ namespace AzureNamingTool.Controllers
         {
             try
             {
-                serviceResponse = await GeneratedNamesService.GetItems();
+                serviceResponse = await _generatedNamesService.GetItemsAsync();
                 if (serviceResponse.Success)
                 {
                     return Ok(serviceResponse.ResponseObject);
@@ -397,7 +417,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -414,7 +434,7 @@ namespace AzureNamingTool.Controllers
         {
             try
             {
-                serviceResponse = await GeneratedNamesService.GetItem(id);
+                serviceResponse = await _generatedNamesService.GetItemAsync(id);
                 if (serviceResponse.Success)
                 {
                     return Ok(serviceResponse.ResponseObject);
@@ -426,7 +446,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -449,14 +469,14 @@ namespace AzureNamingTool.Controllers
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
                         // Get the item details
-                        serviceResponse = await GeneratedNamesService.GetItem(id);
+                        serviceResponse = await _generatedNamesService.GetItemAsync(id);
                         if (serviceResponse.Success)
                         {
                             GeneratedName item = (GeneratedName)serviceResponse.ResponseObject!;
-                            serviceResponse = await GeneratedNamesService.DeleteItem(id);
+                            serviceResponse = await _generatedNamesService.DeleteItemAsync(id);
                             if (serviceResponse.Success)
                             {
-                                AdminLogService.PostItem(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Generated Name (" + item.ResourceName + ") deleted." });
+                                await _adminLogService.PostItemAsync(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Generated Name (" + item.ResourceName + ") deleted." });
                                 CacheHelper.InvalidateCacheObject("GeneratedName");
                                 return Ok("Generated Name (" + item.ResourceName + ") deleted.");
                             }
@@ -483,7 +503,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -502,7 +522,7 @@ namespace AzureNamingTool.Controllers
                 {
                     if (adminpassword == GeneralHelper.DecryptString(config.AdminPassword!, config.SALTKey!))
                     {
-                        serviceResponse = await GeneratedNamesService.DeleteAllItems();
+                        serviceResponse = await _generatedNamesService.DeleteAllItemsAsync();
                         if (serviceResponse.Success)
                         {
                             return Ok(serviceResponse.ResponseObject);
@@ -525,18 +545,18 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
 
         /// <summary>
-        /// This function will reset the site configuration. THIS CANNOT BE UNDONE!
+        /// This function resets the site configuration to the default values.
         /// </summary>
         /// <returns>dttring - Successful operation</returns>
         [HttpPost]
         [Route("[action]")]
-        public IActionResult ResetSiteConfiguration([BindRequired][FromHeader(Name = "AdminPassword")] string adminpassword)
+        public async Task<IActionResult> ResetSiteConfiguration([BindRequired][FromHeader(Name = "AdminPassword")] string adminpassword)
         {
             try
             {
@@ -565,9 +585,11 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
     }
 }
+
+#pragma warning restore CS1591
