@@ -1,12 +1,15 @@
-﻿using AzureNamingTool.Models;
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+using AzureNamingTool.Models;
 using AzureNamingTool.Helpers;
 using Microsoft.AspNetCore.Mvc;
+using Asp.Versioning;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
 using AzureNamingTool.Services;
+using AzureNamingTool.Services.Interfaces;
 using AzureNamingTool.Attributes;
 
 namespace AzureNamingTool.Controllers
@@ -15,10 +18,23 @@ namespace AzureNamingTool.Controllers
     /// Controller for importing and exporting configuration data.
     /// </summary>
     [Route("api/[controller]")]
+    [ApiVersion("1.0")]
     [ApiController]
     [ApiKey]
+    [Produces("application/json")]
     public class ImportExportController : ControllerBase
     {
+        private readonly IImportExportService _importExportService;
+        private readonly IAdminLogService _adminLogService;
+
+        public ImportExportController(
+            IImportExportService importExportService,
+            IAdminLogService adminLogService)
+        {
+            _importExportService = importExportService;
+            _adminLogService = adminLogService;
+        }
+
         /// <summary>
         /// Response for controller functions
         /// </summary>
@@ -36,7 +52,7 @@ namespace AzureNamingTool.Controllers
         {
             try
             {
-                serviceResponse = await ImportExportService.ExportConfig(includeAdmin);
+                serviceResponse = await _importExportService.ExportConfigAsync(includeAdmin);
                 if (serviceResponse.Success)
                 {
                     return Ok(serviceResponse.ResponseObject);
@@ -48,7 +64,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -65,7 +81,7 @@ namespace AzureNamingTool.Controllers
         {
             try
             {
-                serviceResponse = await ImportExportService.PostConfig(configdata);
+                serviceResponse = await _importExportService.PostConfigAsync(configdata);
                 if (serviceResponse.Success)
                 {
                     return Ok(serviceResponse.ResponseObject);
@@ -77,9 +93,11 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
     }
 }
+
+#pragma warning restore CS1591
