@@ -1,5 +1,6 @@
-﻿using AzureNamingTool.Models;
+using AzureNamingTool.Models;
 using Microsoft.AspNetCore.Mvc;
+using Asp.Versioning;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -15,8 +16,10 @@ namespace AzureNamingTool.Controllers
     /// Controller for managing policies.
     /// </summary>
     [Route("api/[controller]")]
+    [ApiVersion("1.0")]
     [ApiController]
     [ApiKey]
+    [Produces("application/json")]
     public class PolicyController : ControllerBase
     {
         //private ServiceResponse serviceResponse = new();
