@@ -1,6 +1,8 @@
-﻿using AzureNamingTool.Models;
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+using AzureNamingTool.Models;
 using AzureNamingTool.Helpers;
 using Microsoft.AspNetCore.Mvc;
+using Asp.Versioning;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,6 +10,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using System.Web;
 using AzureNamingTool.Services;
+using AzureNamingTool.Services.Interfaces;
 using AzureNamingTool.Attributes;
 
 namespace AzureNamingTool.Controllers
@@ -16,10 +19,23 @@ namespace AzureNamingTool.Controllers
     /// Controller for managing resource types.
     /// </summary>
     [Route("api/[controller]")]
+    [ApiVersion("1.0")]
     [ApiController]
     [ApiKey]
+    [Produces("application/json")]
     public class ResourceTypesController : ControllerBase
     {
+        private readonly IResourceTypeService _resourceTypeService;
+        private readonly IAdminLogService _adminLogService;
+
+        public ResourceTypesController(
+            IResourceTypeService resourceTypeService,
+            IAdminLogService adminLogService)
+        {
+            _resourceTypeService = resourceTypeService;
+            _adminLogService = adminLogService;
+        }
+
         private ServiceResponse serviceResponse = new();
 
         // GET: api/<ResourceTypesController>
@@ -29,12 +45,15 @@ namespace AzureNamingTool.Controllers
         /// <param name="admin">bool - Indicates if the user is an admin (optional)</param>
         /// <returns>json - Current resource types data</returns>
         [HttpGet]
+        [ProducesResponseType(typeof(List<Models.ResourceType>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
         public async Task<IActionResult> Get(bool admin = false)
         {
             try
             {
                 // Get list of items
-                serviceResponse = await ResourceTypeService.GetItems(admin);
+                serviceResponse = await _resourceTypeService.GetItemsAsync(admin);
                 if (serviceResponse.Success)
                 {
                     return Ok(serviceResponse.ResponseObject);
@@ -46,7 +65,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -58,12 +77,15 @@ namespace AzureNamingTool.Controllers
         /// <param name="id">int - Resource Type id</param>
         /// <returns>json - Resource Type data</returns>
         [HttpGet("{id}")]
+        [ProducesResponseType(typeof(Models.ResourceType), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
         public async Task<IActionResult> Get(int id)
         {
             try
             {
                 // Get list of items
-                serviceResponse = await ResourceTypeService.GetItem(id);
+                serviceResponse = await _resourceTypeService.GetItemAsync(id);
                 if (serviceResponse.Success)
                 {
                     return Ok(serviceResponse.ResponseObject);
@@ -75,7 +97,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -88,14 +110,17 @@ namespace AzureNamingTool.Controllers
         /// <returns>bool - PASS/FAIL</returns>
         [HttpPost]
         [Route("[action]")]
+        [ProducesResponseType(typeof(bool), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
         public async Task<IActionResult> PostConfig([FromBody] List<ResourceType> items)
         {
             try
             {
-                serviceResponse = await ResourceTypeService.PostConfig(items);
+                serviceResponse = await _resourceTypeService.PostConfigAsync(items);
                 if (serviceResponse.Success)
                 {
-                    AdminLogService.PostItem(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Resource Types updated." });
+                    await _adminLogService.PostItemAsync(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Resource Types updated." });
                     CacheHelper.InvalidateCacheObject("ResourceType");
                     return Ok(serviceResponse.ResponseObject);
                 }
@@ -106,7 +131,7 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
@@ -124,10 +149,10 @@ namespace AzureNamingTool.Controllers
         {
             try
             {
-                serviceResponse = await ResourceTypeService.UpdateTypeComponents(operation, componentid);
+                serviceResponse = await _resourceTypeService.UpdateTypeComponentsAsync(operation, componentid);
                 if (serviceResponse.Success)
                 {
-                    AdminLogService.PostItem(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Resource Types updated." });
+                    await _adminLogService.PostItemAsync(new AdminLogMessage() { Source = "API", Title = "INFORMATION", Message = "Resource Types updated." });
                     CacheHelper.InvalidateCacheObject("ResourceType");
                     return Ok(serviceResponse.ResponseObject);
                 }
@@ -138,9 +163,10 @@ namespace AzureNamingTool.Controllers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 return BadRequest(ex);
             }
         }
     }
 }
+#pragma warning restore CS1591
