@@ -1,5 +1,6 @@
-﻿using AzureNamingTool.Models;
+using AzureNamingTool.Models;
 using AzureNamingTool.Services;
+using AzureNamingTool.Services.Interfaces;
 using Microsoft.AspNetCore.Components.Server.ProtectedBrowserStorage;
 using System;
 
@@ -16,14 +17,15 @@ namespace AzureNamingTool.Helpers
         /// <param name="state">The state container.</param>
         /// <param name="session">The protected session storage.</param>
         /// <param name="name">The username.</param>
+        /// <param name="adminUserService">The admin user service.</param>
         /// <returns>True if the user is an admin user, otherwise false.</returns>
-        public static async Task<bool> IsAdminUser(StateContainer state, ProtectedSessionStorage session, string name)
+        public static async Task<bool> IsAdminUser(StateContainer state, ProtectedSessionStorage session, string name, IAdminUserService adminUserService)
         {
             bool result = false;
             try
             {
                 // Check if the username is in the list of Admin Users
-                ServiceResponse serviceResponse = await AdminUserService.GetItems();
+                ServiceResponse serviceResponse = await adminUserService.GetItemsAsync();
                 if (serviceResponse.Success)
                 {
                     if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
@@ -40,7 +42,8 @@ namespace AzureNamingTool.Helpers
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                // Log error but don't expose to user
+                Console.WriteLine($"Error checking admin user: {ex.Message}");
             }
             return result;
         }
@@ -61,9 +64,8 @@ namespace AzureNamingTool.Helpers
                     currentuser = currentuservalue.Value;
                 }
             }
-            catch (Exception ex)
-            {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+            catch (Exception) {
+                // TODO: Modernize helper - AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
             }
             return currentuser;
         }
