@@ -1,5 +1,7 @@
-﻿using AzureNamingTool.Helpers;
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+using AzureNamingTool.Helpers;
 using AzureNamingTool.Models;
+using AzureNamingTool.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using System;
 
@@ -8,16 +10,22 @@ namespace AzureNamingTool.Services
     /// <summary>
     /// Service for managing the admin settings.
     /// </summary>
-    public class AdminService
+    public class AdminService : IAdminService
     {
+        private readonly IAdminLogService _adminLogService;
         private static readonly SiteConfiguration config = ConfigurationHelper.GetConfigurationData();
+
+        public AdminService(IAdminLogService adminLogService)
+        {
+            _adminLogService = adminLogService;
+        }
 
         /// <summary>
         /// Updates the user password.
         /// </summary>
         /// <param name="password">The new password.</param>
         /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
-        public static async Task<ServiceResponse> UpdatePassword(string password)
+        public async Task<ServiceResponse> UpdatePasswordAsync(string password)
         {
             ServiceResponse serviceResponse = new();
             try
@@ -36,7 +44,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -48,7 +56,7 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="type">The type of API key to generate (fullaccess or readonly).</param>
         /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
-        public static async Task<ServiceResponse> GenerateAPIKey(string type)
+        public async Task<ServiceResponse> GenerateAPIKeyAsync(string type)
         {
             ServiceResponse serviceResponse = new();
             try
@@ -73,7 +81,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -86,7 +94,7 @@ namespace AzureNamingTool.Services
         /// <param name="apikey">The new API key.</param>
         /// <param name="type">The type of API key to update (fullaccess or readonly).</param>
         /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
-        public static async Task<ServiceResponse> UpdateAPIKey(string apikey, string type)
+        public async Task<ServiceResponse> UpdateAPIKeyAsync(string apikey, string type)
         {
             ServiceResponse serviceResponse = new();
             try
@@ -109,7 +117,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -121,7 +129,7 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="identityheadername">The new identity header name.</param>
         /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
-        public static async Task<ServiceResponse> UpdateIdentityHeaderName(string identityheadername)
+        public async Task<ServiceResponse> UpdateIdentityHeaderNameAsync(string identityheadername)
         {
             ServiceResponse serviceResponse = new();
             try
@@ -133,7 +141,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -141,3 +149,4 @@ namespace AzureNamingTool.Services
         }
     }
 }
+#pragma warning restore CS1591
