@@ -1,5 +1,7 @@
-﻿using AzureNamingTool.Helpers;
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+using AzureNamingTool.Helpers;
 using AzureNamingTool.Models;
+using AzureNamingTool.Services.Interfaces;
 using Microsoft.AspNetCore.SignalR;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -9,14 +11,58 @@ namespace AzureNamingTool.Services
     /// <summary>
     /// Service for managing the import and export of configuration data.
     /// </summary>
-    public class ImportExportService
+    public class ImportExportService : IImportExportService
     {
+        private readonly IResourceComponentService _resourceComponentService;
+        private readonly IResourceDelimiterService _resourceDelimiterService;
+        private readonly IResourceEnvironmentService _resourceEnvironmentService;
+        private readonly IResourceFunctionService _resourceFunctionService;
+        private readonly IResourceLocationService _resourceLocationService;
+        private readonly IResourceOrgService _resourceOrgService;
+        private readonly IResourceProjAppSvcService _resourceProjAppSvcService;
+        private readonly IResourceTypeService _resourceTypeService;
+        private readonly IResourceUnitDeptService _resourceUnitDeptService;
+        private readonly ICustomComponentService _customComponentService;
+        private readonly IGeneratedNamesService _generatedNamesService;
+        private readonly IAdminLogService _adminLogService;
+        private readonly IAdminUserService _adminUserService;
+
+        public ImportExportService(
+            IResourceComponentService resourceComponentService,
+            IResourceDelimiterService resourceDelimiterService,
+            IResourceEnvironmentService resourceEnvironmentService,
+            IResourceFunctionService resourceFunctionService,
+            IResourceLocationService resourceLocationService,
+            IResourceOrgService resourceOrgService,
+            IResourceProjAppSvcService resourceProjAppSvcService,
+            IResourceTypeService resourceTypeService,
+            IResourceUnitDeptService resourceUnitDeptService,
+            ICustomComponentService customComponentService,
+            IGeneratedNamesService generatedNamesService,
+            IAdminLogService adminLogService,
+            IAdminUserService adminUserService)
+        {
+            _resourceComponentService = resourceComponentService;
+            _resourceDelimiterService = resourceDelimiterService;
+            _resourceEnvironmentService = resourceEnvironmentService;
+            _resourceFunctionService = resourceFunctionService;
+            _resourceLocationService = resourceLocationService;
+            _resourceOrgService = resourceOrgService;
+            _resourceProjAppSvcService = resourceProjAppSvcService;
+            _resourceTypeService = resourceTypeService;
+            _resourceUnitDeptService = resourceUnitDeptService;
+            _customComponentService = customComponentService;
+            _generatedNamesService = generatedNamesService;
+            _adminLogService = adminLogService;
+            _adminUserService = adminUserService;
+        }
+
         /// <summary>
         /// Export the configuration data.
         /// </summary>
         /// <param name="includeadmin">Flag to include admin settings in the export.</param>
         /// <returns>The service response containing the exported configuration data.</returns>
-        public static async Task<ServiceResponse> ExportConfig(bool includeadmin = false)
+        public async Task<ServiceResponse> ExportConfigAsync(bool includeadmin = false)
         {
             ServiceResponse serviceResponse = new();
             try
@@ -24,7 +70,7 @@ namespace AzureNamingTool.Services
                 ConfigurationData configdata = new();
                 // Get the current data
                 //ResourceComponents
-                serviceResponse = await ResourceComponentService.GetItems(true);
+                serviceResponse = await _resourceComponentService.GetItemsAsync(true);
                 if (serviceResponse.Success)
                 {
                     if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
@@ -34,77 +80,77 @@ namespace AzureNamingTool.Services
                 }
 
                 //ResourceDelimiters
-                serviceResponse = await ResourceDelimiterService.GetItems(true);
+                serviceResponse = await _resourceDelimiterService.GetItemsAsync(true);
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceDelimiters = serviceResponse.ResponseObject!;
                 }
 
                 //ResourceEnvironments
-                serviceResponse = await ResourceEnvironmentService.GetItems();
+                serviceResponse = await _resourceEnvironmentService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceEnvironments = serviceResponse.ResponseObject!;
                 }
 
                 // ResourceFunctions
-                serviceResponse = await ResourceFunctionService.GetItems();
+                serviceResponse = await _resourceFunctionService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceFunctions = serviceResponse.ResponseObject!;
                 }
 
                 // ResourceLocations
-                serviceResponse = await ResourceLocationService.GetItems();
+                serviceResponse = await _resourceLocationService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceLocations = serviceResponse.ResponseObject!;
                 }
 
                 // ResourceOrgs
-                serviceResponse = await ResourceOrgService.GetItems();
+                serviceResponse = await _resourceOrgService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceOrgs = serviceResponse.ResponseObject!;
                 }
 
                 // ResourceProjAppSvc
-                serviceResponse = await ResourceProjAppSvcService.GetItems();
+                serviceResponse = await _resourceProjAppSvcService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceProjAppSvcs = serviceResponse.ResponseObject!;
                 }
 
                 // ResourceTypes
-                serviceResponse = await ResourceTypeService.GetItems();
+                serviceResponse = await _resourceTypeService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceTypes = serviceResponse.ResponseObject!;
                 }
 
                 // ResourceUnitDepts
-                serviceResponse = await ResourceUnitDeptService.GetItems();
+                serviceResponse = await _resourceUnitDeptService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.ResourceUnitDepts = serviceResponse.ResponseObject!;
                 }
 
                 // CustomComponents
-                serviceResponse = await CustomComponentService.GetItems();
+                serviceResponse = await _customComponentService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.CustomComponents = serviceResponse.ResponseObject!;
                 }
 
                 //GeneratedNames
-                serviceResponse = await GeneratedNamesService.GetItems();
+                serviceResponse = await _generatedNamesService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.GeneratedNames = serviceResponse.ResponseObject!;
                 }
 
                 //AdminLogs
-                serviceResponse = await AdminLogService.GetItems();
+                serviceResponse = await _adminLogService.GetItemsAsync();
                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                 {
                     configdata.AdminLogs = serviceResponse.ResponseObject;
@@ -140,7 +186,7 @@ namespace AzureNamingTool.Services
                     //IdentityHeaderName
                     configdata.IdentityHeaderName = config.IdentityHeaderName;
                     //AdminUsers
-                    serviceResponse = await AdminUserService.GetItems();
+                    serviceResponse = await _adminUserService.GetItemsAsync();
                     if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                     {
                         configdata.AdminUsers = serviceResponse.ResponseObject!;
@@ -152,7 +198,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -164,30 +210,30 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="configdata">The configuration data to import.</param>
         /// <returns>The service response indicating the success or failure of the import.</returns>
-        public static async Task<ServiceResponse> PostConfig(ConfigurationData configdata)
+        public async Task<ServiceResponse> PostConfigAsync(ConfigurationData configdata)
         {
             ServiceResponse serviceResponse = new();
             try
             {
                 // Write all the configurations
-                await ResourceComponentService.PostConfig(configdata.ResourceComponents);
-                await ResourceDelimiterService.PostConfig(configdata.ResourceDelimiters);
-                await ResourceEnvironmentService.PostConfig(configdata.ResourceEnvironments);
-                await ResourceFunctionService.PostConfig(configdata.ResourceFunctions);
-                await ResourceLocationService.PostConfig(configdata.ResourceLocations);
-                await ResourceOrgService.PostConfig(configdata.ResourceOrgs);
-                await ResourceProjAppSvcService.PostConfig(configdata.ResourceProjAppSvcs);
-                await ResourceTypeService.PostConfig(configdata.ResourceTypes);
-                await ResourceUnitDeptService.PostConfig(configdata.ResourceUnitDepts);
-                await CustomComponentService.PostConfig(configdata.CustomComponents);
-                await GeneratedNamesService.PostConfig(configdata.GeneratedNames);
+                await _resourceComponentService.PostConfigAsync(configdata.ResourceComponents);
+                await _resourceDelimiterService.PostConfigAsync(configdata.ResourceDelimiters);
+                await _resourceEnvironmentService.PostConfigAsync(configdata.ResourceEnvironments);
+                await _resourceFunctionService.PostConfigAsync(configdata.ResourceFunctions);
+                await _resourceLocationService.PostConfigAsync(configdata.ResourceLocations);
+                await _resourceOrgService.PostConfigAsync(configdata.ResourceOrgs);
+                await _resourceProjAppSvcService.PostConfigAsync(configdata.ResourceProjAppSvcs);
+                await _resourceTypeService.PostConfigAsync(configdata.ResourceTypes);
+                await _resourceUnitDeptService.PostConfigAsync(configdata.ResourceUnitDepts);
+                await _customComponentService.PostConfigAsync(configdata.CustomComponents);
+                await _generatedNamesService.PostConfigAsync(configdata.GeneratedNames);
                 if (GeneralHelper.IsNotNull(configdata.AdminUsers))
                 {
-                    await AdminUserService.PostConfig(configdata.AdminUsers);
+                    await _adminUserService.PostConfigAsync(configdata.AdminUsers);
                 }
                 if (GeneralHelper.IsNotNull(configdata.AdminLogs))
                 {
-                    await AdminLogService.PostConfig(configdata.AdminLogs);
+                    await _adminLogService.PostConfigAsync(configdata.AdminLogs);
                 }
 
                 var config = ConfigurationHelper.GetConfigurationData();
@@ -283,7 +329,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -291,3 +337,5 @@ namespace AzureNamingTool.Services
         }
     }
 }
+
+#pragma warning restore CS1591
