@@ -1,5 +1,9 @@
-﻿using AzureNamingTool.Helpers;
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+using AzureNamingTool.Helpers;
 using AzureNamingTool.Models;
+using AzureNamingTool.Repositories;
+using AzureNamingTool.Repositories.Interfaces;
+using AzureNamingTool.Services.Interfaces;
 using System.Net.WebSockets;
 using System.Security.AccessControl;
 
@@ -8,20 +12,33 @@ namespace AzureNamingTool.Services
     /// <summary>
     /// Service for managing resource components.
     /// </summary>
-    public class ResourceComponentService
+    public class ResourceComponentService : IResourceComponentService
     {
+        private readonly IConfigurationRepository<ResourceComponent> _repository;
+        private readonly IAdminLogService _adminLogService;
+        private readonly IResourceConfigurationCoordinator _coordinator;
+
+        public ResourceComponentService(
+            IConfigurationRepository<ResourceComponent> repository,
+            IAdminLogService adminLogService,
+            IResourceConfigurationCoordinator coordinator)
+        {
+            _repository = repository;
+            _adminLogService = adminLogService;
+            _coordinator = coordinator;
+        }
 
         /// <summary>
         /// Retrieves a list of resource components.
         /// </summary>
         /// <param name="admin">A boolean value indicating whether the user is an admin.</param>
         /// <returns>A service response containing the list of resource components.</returns>
-        public static async Task<ServiceResponse> GetItems(bool admin)
+        public async Task<ServiceResponse> GetItemsAsync(bool admin = true)
         {
             ServiceResponse serviceResponse = new();
             try
             {
-                var items = await ConfigurationHelper.GetList<ResourceComponent>();
+                var items = (await _repository.GetAllAsync()).ToList();
                 if (GeneralHelper.IsNotNull(items))
                 {
                     if (!admin)
@@ -30,7 +47,8 @@ namespace AzureNamingTool.Services
                     }
                     else
                     {
-                        serviceResponse.ResponseObject = items.OrderBy(y => y.SortOrder).OrderByDescending(y => y.Enabled).ToList();
+                        // Admin view: show all items in their exact SortOrder
+                        serviceResponse.ResponseObject = items.OrderBy(y => y.SortOrder).ToList();
                     }
                     serviceResponse.Success = true;
                 }
@@ -41,7 +59,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -53,13 +71,13 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="id">The ID of the resource component item.</param>
         /// <returns>A service response containing the resource component item.</returns>
-        public static async Task<ServiceResponse> GetItem(int id)
+        public async Task<ServiceResponse> GetItemAsync(int id)
         {
             ServiceResponse serviceResponse = new();
             try
             {
                 // Get list of items
-                var items = await ConfigurationHelper.GetList<ResourceComponent>();
+                var items = (await _repository.GetAllAsync()).ToList();
                 if (GeneralHelper.IsNotNull(items))
                 {
                     var item = items.Find(x => x.Id == id);
@@ -80,7 +98,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -92,13 +110,13 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="item">The resource component item to post.</param>
         /// <returns>A service response indicating the success of the operation.</returns>
-        public static async Task<ServiceResponse> PostItem(ResourceComponent item)
+        public async Task<ServiceResponse> PostItemAsync(ResourceComponent item)
         {
             ServiceResponse serviceResponse = new();
             try
             {
                 // Get list of items
-                var items = await ConfigurationHelper.GetList<ResourceComponent>();
+                var items = (await _repository.GetAllAsync()).ToList();
                 if (GeneralHelper.IsNotNull(items))
                 {
                     if (GeneralHelper.IsNotNull(items))
@@ -133,7 +151,7 @@ namespace AzureNamingTool.Services
                             }
 
                             // Reset the sort order of the list
-                            foreach (ResourceComponent thisitem in items.OrderBy(x => x.SortOrder).OrderByDescending(x => x.Enabled).ToList())
+                            foreach (ResourceComponent thisitem in items.OrderBy(x => x.SortOrder).ToList())
                             {
                                 thisitem.SortOrder = position;
                                 position += 1;
@@ -159,25 +177,26 @@ namespace AzureNamingTool.Services
                         }
 
                         position = 1;
-                        foreach (ResourceComponent thisitem in items.OrderBy(x => x.SortOrder).OrderByDescending(x => x.Enabled).ToList())
+                        foreach (ResourceComponent thisitem in items.OrderBy(x => x.SortOrder).ToList())
                         {
                             thisitem.SortOrder = position;
-                            thisitem.Id = position;
+                            // DO NOT reassign Id - it should remain stable!
+                            // thisitem.Id = position;
                             position += 1;
                         }
 
                         // Write items to file
-                        await ConfigurationHelper.WriteList<ResourceComponent>(items);
+                        await _repository.SaveAllAsync(items);
                         serviceResponse.Success = true;
                         // Get the item
-                        var newitem = (await ResourceComponentService.GetItem((int)item.Id)).ResponseObject;
+                        var newitem = (await GetItemAsync((int)item.Id)).ResponseObject;
                         serviceResponse.ResponseObject = newitem;
                     }
                 }
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -189,73 +208,41 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="id">The ID of the item to delete.</param>
         /// <returns>A service response indicating the success of the operation.</returns>
-        public static async Task<ServiceResponse> DeleteItem(int id)
+        public async Task<ServiceResponse> DeleteItemAsync(int id)
         {
             ServiceResponse serviceResponse = new();
             try
             {
                 // Get list of items
-                var items = await ConfigurationHelper.GetList<ResourceComponent>();
+                var items = (await _repository.GetAllAsync()).ToList();
                 if (GeneralHelper.IsNotNull(items))
                 {
                     // Get the specified item
                     var item = items.Find(x => x.Id == id && x.IsCustom == true);
                     if (GeneralHelper.IsNotNull(item))
                     {
-                        // Delete any resource type settings for the component
-                        List<string> currentvalues = [];
-                        serviceResponse = await ResourceTypeService.GetItems();
-                        if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
+                        // Delete any resource type settings for the component using coordinator
+                        await _coordinator.UpdateTypesOnComponentDeleteAsync(item.Name);
+
+                        // Delete any custom components for this resource component using coordinator
+                        await _coordinator.DeleteCustomComponentsByParentIdAsync(id);
+
+                        // Remove the item from the collection
+                        items.Remove(item);
+
+                        // Update all the sort order values to reflect the removal
+                        int position = 1;
+                        foreach (ResourceComponent thisitem in items.OrderBy(x => x.SortOrder).ToList())
                         {
-                            List<Models.ResourceType> resourceTypes = (List<Models.ResourceType>)serviceResponse.ResponseObject!;
-                            if (GeneralHelper.IsNotNull(resourceTypes))
-                            {
-                                foreach (Models.ResourceType currenttype in resourceTypes)
-                                {
-                                    currentvalues = new List<string>(currenttype.Optional.Split(','));
-                                    if (currentvalues.Contains(GeneralHelper.NormalizeName(item.Name, false)))
-                                    {
-                                        currentvalues.Remove(GeneralHelper.NormalizeName(item.Name, false));
-                                        currenttype.Optional = String.Join(",", [.. currentvalues]);
-                                    }
-
-                                    currentvalues = new List<string>(currenttype.Exclude.Split(','));
-                                    if (currentvalues.Contains(GeneralHelper.NormalizeName(item.Name, false)))
-                                    {
-                                        currentvalues.Remove(GeneralHelper.NormalizeName(item.Name, false));
-                                        currenttype.Exclude = String.Join(",", [.. currentvalues]);
-                                    }
-                                    await ResourceTypeService.PostItem(currenttype);
-                                }
-
-                                // Delete any custom components for this resource component
-                                await CustomComponentService.DeleteByParentComponentId(id);
-
-                                // Remove the item from the collection
-                                items.Remove(item);
-
-                                // Update all the sort order values to reflect the removal
-                                int position = 1;
-                                foreach (ResourceComponent thisitem in items.OrderBy(x => x.SortOrder).ToList())
-                                {
-                                    thisitem.SortOrder = position;
-                                    thisitem.Id = position;
-                                    position += 1;
-                                }
-
-                                // Write items to file
-                                await ConfigurationHelper.WriteList<ResourceComponent>(items);
-                                serviceResponse.Success = true;
-                                if (!serviceResponse.Success)
-                                { 
-                                    serviceResponse.ResponseObject = "Custom Component deletion failed! Please check the Admin log for details.";
-                                }
-                            }
-                            else
-                            {
-                                serviceResponse.ResponseObject = "Resource Types not found!";
-                            }
+                            thisitem.SortOrder = position;
+                            // DO NOT reassign Id - it should remain stable!
+                            // thisitem.Id = position;
+                            position += 1;
                         }
+
+                        // Write items to file
+                        await _repository.SaveAllAsync(items);
+                        serviceResponse.Success = true;
                     }
                     else
                     {
@@ -269,7 +256,7 @@ namespace AzureNamingTool.Services
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.ResponseObject = ex;
                 serviceResponse.Success = false;
             }
@@ -281,7 +268,7 @@ namespace AzureNamingTool.Services
         /// </summary>
         /// <param name="items">The list of resource components to configure.</param>
         /// <returns>A service response indicating the success of the operation.</returns>
-        public static async Task<ServiceResponse> PostConfig(List<ResourceComponent> items)
+        public async Task<ServiceResponse> PostConfigAsync(List<ResourceComponent> items)
         {
             ServiceResponse serviceResponse = new();
             try
@@ -316,23 +303,52 @@ namespace AzureNamingTool.Services
                     }
                 }
 
-                // Determine new item ids
-                int i = 1;
+                // Determine new item ids (keep existing IDs, only assign to new items)
+                long maxId = newitems.Any() ? newitems.Max(x => x.Id) : 0;
 
-                foreach (ResourceComponent item in newitems.OrderByDescending(x => x.Enabled).OrderBy(x => x.SortOrder))
+                foreach (ResourceComponent item in newitems)
                 {
-                    item.Id = i;
-                    item.SortOrder = i;
-                    i += 1;
+                    // Only assign new IDs to items that don't have one
+                    if (item.Id == 0)
+                    {
+                        maxId++;
+                        item.Id = maxId;
+                    }
+                    
+                    // Keep the existing SortOrder - don't modify it!
                 }
 
                 // Write items to file
-                await ConfigurationHelper.WriteList<ResourceComponent>(newitems);
+                await _repository.SaveAllAsync(newitems);
                 serviceResponse.Success = true;
             }
             catch (Exception ex)
             {
-                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                serviceResponse.Success = false;
+                serviceResponse.ResponseObject = ex;
+            }
+            return serviceResponse;
+        }
+
+        /// <summary>
+        /// Updates the sort order of resource components without normalization.
+        /// This method directly saves the provided items without re-sequencing IDs or sort orders.
+        /// </summary>
+        /// <param name="items">The list of resource components with updated sort orders.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
+        public async Task<ServiceResponse> UpdateSortOrderAsync(List<ResourceComponent> items)
+        {
+            ServiceResponse serviceResponse = new();
+            try
+            {
+                // Simply save the items as-is without any normalization
+                await _repository.SaveAllAsync(items);
+                serviceResponse.Success = true;
+            }
+            catch (Exception ex)
+            {
+                await _adminLogService.PostItemAsync(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
                 serviceResponse.Success = false;
                 serviceResponse.ResponseObject = ex;
             }
@@ -340,3 +356,5 @@ namespace AzureNamingTool.Services
         }
     }
 }
+
+#pragma warning restore CS1591
